get_groups

returns a list of groups for a multi group model

Usage

get_groups(multi_group_model)

Arguments

multi_group_model

multi group model created with mxsem_group_by

Value

list with data for each group

Examples

# THE FOLLOWING EXAMPLE IS ADAPTED FROM
# https://openmx.ssri.psu.edu/docs/OpenMx/latest/_static/Rdoc/mxModel.html
library(mxsem)

model <- 'spatial =~ visual + cubes + paper
          verbal  =~ general + paragrap + sentence
          math    =~ numeric + series + arithmet'

mg_model <- mxsem(model = model,
                  data  = OpenMx::HS.ability.data) |>
  # we want separate models for all combinations of grades and schools:
  mxsem_group_by(grouping_variables = "school") |>
  mxTryHard()
#> Warning: This function is very experimental and may not yet work properly. Use with caution.
#> The following parameters will be the same across groups: 
#> The following parameters will be group specific: spatial→cubes, spatial→paper, verbal→paragrap, verbal→sentence, math→series, math→arithmet, visual↔visual, cubes↔cubes, paper↔paper, general↔general, paragrap↔paragrap, sentence↔sentence, numeric↔numeric, series↔series, arithmet↔arithmet, spatial↔spatial, spatial↔verbal, verbal↔verbal, spatial↔math, verbal↔math, math↔math, one→visual, one→cubes, one→paper, one→general, one→paragrap, one→sentence, one→numeric, one→series, one→arithmet
#> Running untitled1 with 60 parameters
#> 
#> Beginning initial fit attempt
#> Running untitled1 with 60 parameters
#> 
#>  Lowest minimum so far:  15969.1164364042
#>  Not all eigenvalues of the Hessian are positive: 402219341.891002, 24136125.6511393, 13629.6144673512, 13392.3232184624, 5863.79847475932, 4759.40902391292, 4360.27273268796, 1952.6181949062, 1825.43588373687, 1316.14320026041, 1149.69073403232, 357.020859626818, 139.587433865707, 71.9069260853621, 69.9404892014232, 46.269886249906, 42.6674235408757, 38.5891239225252, 24.7840310381383, 23.5787338184784, 22.843212497112, 21.5482242637395, 18.3282408584932, 15.9762169652866, 15.3060860161677, 13.9445602579177, 9.90204880539195, 6.10860281338663, 5.88811684225566, 5.50821258581185, 4.37657765716972, 4.36058933723348, 3.16509425753877, 2.38968859583729, 1.63544781213769, 1.58365930943974, 1.58124508589757, 1.38618515635273, 1.05930596548883, 0.739622567890229, 0.519989228275848, 0.479705953325449, 0.408746816359721, 0.367294348213997, 0.32782118531774, 0.18308465960063, 0.152239339155351, 0.0815597420870642, 0.0637842634380338, 0.0438074692131403, 0.0383380250844334, 0.0370122618603593, 0.0294527879632414, 0.0209545898869084, 0.0194967028842953, 0.00904037989428845, 0.00750007582366306, 0.00577013380997979, 2.1042732070824e-05, -0.000242048900633485
#> 
#> Beginning fit attempt 1 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#>  Not all eigenvalues of the Hessian are positive: 402219341.891002, 24136125.6511393, 13629.6144673512, 13392.3232184624, 5863.79847475932, 4759.40902391292, 4360.27273268796, 1952.6181949062, 1825.43588373687, 1316.14320026041, 1149.69073403232, 357.020859626818, 139.587433865707, 71.9069260853621, 69.9404892014232, 46.269886249906, 42.6674235408757, 38.5891239225252, 24.7840310381383, 23.5787338184784, 22.843212497112, 21.5482242637395, 18.3282408584932, 15.9762169652866, 15.3060860161677, 13.9445602579177, 9.90204880539195, 6.10860281338663, 5.88811684225566, 5.50821258581185, 4.37657765716972, 4.36058933723348, 3.16509425753877, 2.38968859583729, 1.63544781213769, 1.58365930943974, 1.58124508589757, 1.38618515635273, 1.05930596548883, 0.739622567890229, 0.519989228275848, 0.479705953325449, 0.408746816359721, 0.367294348213997, 0.32782118531774, 0.18308465960063, 0.152239339155351, 0.0815597420870642, 0.0637842634380338, 0.0438074692131403, 0.0383380250844334, 0.0370122618603593, 0.0294527879632414, 0.0209545898869084, 0.0194967028842953, 0.00904037989428845, 0.00750007582366306, 0.00577013380997979, 2.1042732070824e-05, -0.000242048900633485
#> 
#> Beginning fit attempt 2 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Fit attempt generated errors
#> 
#> Beginning fit attempt 3 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Fit attempt generated errors
#> 
#> Beginning fit attempt 4 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Fit attempt generated errors
#> 
#> Beginning fit attempt 5 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Fit attempt worse than current best:  16209.6871216905 vs 15969.1164364042
#> 
#> Beginning fit attempt 6 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Fit attempt generated errors
#> 
#> Beginning fit attempt 7 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Lowest minimum so far:  15964.3086826835
#>  OpenMx status code 6 not in list of acceptable status codes, (0,0)
#> 
#> Beginning fit attempt 8 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Lowest minimum so far:  15915.8867825839
#>  Not all eigenvalues of the Hessian are positive: 2990849397.09345, 36679.7130742452, 12240.4646633526, 5730.88779967562, 4937.3106071697, 1953.75503648136, 1890.3254289246, 1142.34456213486, 787.04790598729, 354.396995187624, 310.047197250154, 140.305073518561, 71.9584036484525, 69.7746432452137, 46.2894876787585, 44.9972763095927, 42.0086317582415, 24.9074887820622, 23.5220184318479, 22.8583992988996, 20.9236069173795, 18.3029411074959, 16.0859169153269, 16.0186320628057, 15.3106454650595, 9.90788680821519, 8.4536562041252, 5.88883164182815, 5.53397535358928, 4.35911437748761, 4.18469435406321, 3.49078254558558, 3.16679218639257, 2.82748599177605, 1.59877573155292, 1.51491589978386, 1.4396850087504, 1.1584010358673, 1.0550578239879, 0.739843427161077, 0.518696614914753, 0.503028837426428, 0.492307724150896, 0.411139044433053, 0.327916137906301, 0.240805552974001, 0.184223460882813, 0.150794726394433, 0.0854641889241054, 0.0835849196028065, 0.041455365312845, 0.0385750185508065, 0.0369547752719925, 0.0340109673066367, 0.0295555712090366, 0.0209735081745827, 0.00783733046407599, 0.00757126204632692, 0.00246501781056233, -1.88357398931716e-06
#> 
#> Beginning fit attempt 9 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Lowest minimum so far:  15915.8867825839
#>  Not all eigenvalues of the Hessian are positive: 2990849397.14111, 36679.7092137756, 12240.4942540492, 5730.93458976095, 4937.31261827489, 1953.77659601973, 1890.32652242487, 1142.3575097611, 787.048138490903, 354.412245693047, 310.080943139797, 140.303575586627, 71.9586913667582, 69.7746432487314, 46.2895699040864, 44.997276542044, 42.0079561026726, 24.9074887926374, 23.522109876087, 22.8584820096878, 20.9236069329989, 18.3029410991335, 16.085916943336, 16.0187188745163, 15.3106413198505, 9.90796294736151, 8.45365620521807, 5.89109694505256, 5.53502770022363, 4.35917507999673, 4.18469430612558, 3.49040983730769, 3.16682117743491, 2.82731627121386, 1.59866387443042, 1.51584896551967, 1.4396846416934, 1.15842083370562, 1.05554073654078, 0.740010679246952, 0.518655314536687, 0.502981577291251, 0.492432654285984, 0.411224136272039, 0.32798249285585, 0.240815820319593, 0.184304789285742, 0.150859640209499, 0.0854321576259218, 0.0836205509414968, 0.0414693961745321, 0.0385844638621818, 0.0371642731016848, 0.0339869391900638, 0.0295576218433882, 0.0210917228739541, 0.0078480988479631, 0.00757444530126258, 0.0024877350474363, -1.82083468725078e-06
#> 
#> Beginning fit attempt 10 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Lowest minimum so far:  15915.8848704949
#> 
#> Retry limit reached
#> 
#> Solution found
#> 
#> 
#>  Solution found!  Final fit=15915.885 (started at 11279122)  (11 attempt(s): 7 valid, 4 errors)
#>  Start values from best fit:
#> 0.42948198565436,0.285882523557041,0.31187316113002,0.441812762186046,2.27731454284791,0.956809987467692,23.8755567975854,15.2674767631162,5.97882306529234,55.1061631400964,3.53281685217934,5.84888211392766,14.0564590978876,39.6456245639448,13.7019343190686,23.6205401720743,25.2466602567814,80.1283852688529,12.1228298080201,19.0151822013261,9.14078194150063,29.5793206055627,24.8000163770941,14.3034619394874,44.8484070879003,9.95174912294824,18.8483064487107,14.4551881731066,18.7517241861414,25.8276061407374,0.31840806107535,0.222398073994069,0.302390397625809,0.479319509302114,-470.481334667538,-135.563781233199,17.459455047776,20.7375800626593,6.42403274604368,48.0423730609026,3.66224077348386,6.68651219635686,19.4751966469782,12.3901799625083,13.7546849588173,32.7918389613925,25.3404547030945,89.8641441587419,-0.0668942532343407,-0.0843635724569729,0.000295451308814675,29.6474499699091,23.9359165924698,14.1602738612667,36.6987959121337,8.46796566015973,15.9807993148802,13.891021342072,17.5448462898677,22.769230577266

# let's summarize the results:
summarize_multi_group_model(mg_model)
#> 
#> 
#> --------------------------------------------------------------------------------
#> Common Parameters:
#>  [1] name      matrix    row       col       Estimate  Std.Error lbound   
#>  [8] ubound    lboundMet uboundMet
#> <0 rows> (or 0-length row.names)
#> --------------------------------------------------------------------------------
#> group_1:
#> # A tibble: 1 × 1
#>   school     
#>   <fct>      
#> 1 Grant-White
#> 
#>                         name              matrix      row      col   Estimate
#> 1      spatial→cubes_group_1 untitled1_group_1.A    cubes  spatial  0.4294820
#> 2      spatial→paper_group_1 untitled1_group_1.A    paper  spatial  0.2858825
#> 3    verbal→paragrap_group_1 untitled1_group_1.A paragrap   verbal  0.3118732
#> 4    verbal→sentence_group_1 untitled1_group_1.A sentence   verbal  0.4418128
#> 5        math→series_group_1 untitled1_group_1.A   series     math  2.2773145
#> 6      math→arithmet_group_1 untitled1_group_1.A arithmet     math  0.9568100
#> 7      visual↔visual_group_1 untitled1_group_1.S   visual   visual 23.8755568
#> 8        cubes↔cubes_group_1 untitled1_group_1.S    cubes    cubes 15.2674768
#> 9        paper↔paper_group_1 untitled1_group_1.S    paper    paper  5.9788231
#> 10   general↔general_group_1 untitled1_group_1.S  general  general 55.1061631
#> 11 paragrap↔paragrap_group_1 untitled1_group_1.S paragrap paragrap  3.5328169
#> 12 sentence↔sentence_group_1 untitled1_group_1.S sentence sentence  5.8488821
#> 13   numeric↔numeric_group_1 untitled1_group_1.S  numeric  numeric 14.0564591
#> 14     series↔series_group_1 untitled1_group_1.S   series   series 39.6456246
#> 15 arithmet↔arithmet_group_1 untitled1_group_1.S arithmet arithmet 13.7019343
#> 16   spatial↔spatial_group_1 untitled1_group_1.S  spatial  spatial 23.6205402
#> 17    spatial↔verbal_group_1 untitled1_group_1.S  spatial   verbal 25.2466603
#> 18     verbal↔verbal_group_1 untitled1_group_1.S   verbal   verbal 80.1283853
#> 19      spatial↔math_group_1 untitled1_group_1.S  spatial     math 12.1228298
#> 20       verbal↔math_group_1 untitled1_group_1.S   verbal     math 19.0151822
#> 21         math↔math_group_1 untitled1_group_1.S     math     math  9.1407819
#> 22        one→visual_group_1 untitled1_group_1.M        1   visual 29.5793206
#> 23         one→cubes_group_1 untitled1_group_1.M        1    cubes 24.8000164
#> 24         one→paper_group_1 untitled1_group_1.M        1    paper 14.3034619
#> 25       one→general_group_1 untitled1_group_1.M        1  general 44.8484071
#> 26      one→paragrap_group_1 untitled1_group_1.M        1 paragrap  9.9517491
#> 27      one→sentence_group_1 untitled1_group_1.M        1 sentence 18.8483064
#> 28       one→numeric_group_1 untitled1_group_1.M        1  numeric 14.4551882
#> 29        one→series_group_1 untitled1_group_1.M        1   series 18.7517242
#> 30      one→arithmet_group_1 untitled1_group_1.M        1 arithmet 25.8276061
#>     Std.Error lbound ubound lboundMet uboundMet
#> 1  24.9173937     NA     NA     FALSE     FALSE
#> 2  39.6842595     NA     NA     FALSE     FALSE
#> 3  93.5308040     NA     NA     FALSE     FALSE
#> 4  70.2430034     NA     NA     FALSE     FALSE
#> 5   9.3356522     NA     NA     FALSE     FALSE
#> 6  16.7180950     NA     NA     FALSE     FALSE
#> 7   0.5214630  1e-06     NA     FALSE     FALSE
#> 8   1.0289503  1e-06     NA     FALSE     FALSE
#> 9   2.5983822  1e-06     NA     FALSE     FALSE
#> 10  0.2524517  1e-06     NA     FALSE     FALSE
#> 11  3.4806967  1e-06     NA     FALSE     FALSE
#> 12  1.9305389  1e-06     NA     FALSE     FALSE
#> 13  1.0511771  1e-06     NA     FALSE     FALSE
#> 14  0.3252276  1e-06     NA     FALSE     FALSE
#> 15  1.0882715  1e-06     NA     FALSE     FALSE
#> 16  0.6511978  1e-06     NA     FALSE     FALSE
#> 17  0.6036360     NA     NA     FALSE     FALSE
#> 18  0.2714431  1e-06     NA     FALSE     FALSE
#> 19  1.8690565     NA     NA     FALSE     FALSE
#> 20  1.2101676     NA     NA     FALSE     FALSE
#> 21  2.2911261  1e-06     NA     FALSE     FALSE
#> 22  4.2136530     NA     NA     FALSE     FALSE
#> 23  5.9194746     NA     NA     FALSE     FALSE
#> 24  9.4068284     NA     NA     FALSE     FALSE
#> 25  2.9320130     NA     NA     FALSE     FALSE
#> 26 10.8874767     NA     NA     FALSE     FALSE
#> 27  8.1085918     NA     NA     FALSE     FALSE
#> 28  5.9839044     NA     NA     FALSE     FALSE
#> 29  3.3276581     NA     NA     FALSE     FALSE
#> 30  6.0888489     NA     NA     FALSE     FALSE
#> --------------------------------------------------------------------------------
#> group_2:
#> # A tibble: 1 × 1
#>   school 
#>   <fct>  
#> 1 Pasteur
#> 
#>                         name              matrix      row      col
#> 31     spatial→cubes_group_2 untitled1_group_2.A    cubes  spatial
#> 32     spatial→paper_group_2 untitled1_group_2.A    paper  spatial
#> 33   verbal→paragrap_group_2 untitled1_group_2.A paragrap   verbal
#> 34   verbal→sentence_group_2 untitled1_group_2.A sentence   verbal
#> 35       math→series_group_2 untitled1_group_2.A   series     math
#> 36     math→arithmet_group_2 untitled1_group_2.A arithmet     math
#> 37     visual↔visual_group_2 untitled1_group_2.S   visual   visual
#> 38       cubes↔cubes_group_2 untitled1_group_2.S    cubes    cubes
#> 39       paper↔paper_group_2 untitled1_group_2.S    paper    paper
#> 40   general↔general_group_2 untitled1_group_2.S  general  general
#> 41 paragrap↔paragrap_group_2 untitled1_group_2.S paragrap paragrap
#> 42 sentence↔sentence_group_2 untitled1_group_2.S sentence sentence
#> 43   numeric↔numeric_group_2 untitled1_group_2.S  numeric  numeric
#> 44     series↔series_group_2 untitled1_group_2.S   series   series
#> 45 arithmet↔arithmet_group_2 untitled1_group_2.S arithmet arithmet
#> 46   spatial↔spatial_group_2 untitled1_group_2.S  spatial  spatial
#> 47    spatial↔verbal_group_2 untitled1_group_2.S  spatial   verbal
#> 48     verbal↔verbal_group_2 untitled1_group_2.S   verbal   verbal
#> 49      spatial↔math_group_2 untitled1_group_2.S  spatial     math
#> 50       verbal↔math_group_2 untitled1_group_2.S   verbal     math
#> 51         math↔math_group_2 untitled1_group_2.S     math     math
#> 52        one→visual_group_2 untitled1_group_2.M        1   visual
#> 53         one→cubes_group_2 untitled1_group_2.M        1    cubes
#> 54         one→paper_group_2 untitled1_group_2.M        1    paper
#> 55       one→general_group_2 untitled1_group_2.M        1  general
#> 56      one→paragrap_group_2 untitled1_group_2.M        1 paragrap
#> 57      one→sentence_group_2 untitled1_group_2.M        1 sentence
#> 58       one→numeric_group_2 untitled1_group_2.M        1  numeric
#> 59        one→series_group_2 untitled1_group_2.M        1   series
#> 60      one→arithmet_group_2 untitled1_group_2.M        1 arithmet
#>         Estimate    Std.Error lbound ubound lboundMet uboundMet
#> 31  3.184081e-01 2.685735e+01     NA     NA     FALSE     FALSE
#> 32  2.223981e-01 4.787273e+01     NA     NA     FALSE     FALSE
#> 33  3.023904e-01 1.021608e+02     NA     NA     FALSE     FALSE
#> 34  4.793195e-01 7.128747e+01     NA     NA     FALSE     FALSE
#> 35 -4.704813e+02 7.138093e-02     NA     NA     FALSE     FALSE
#> 36 -1.355638e+02 1.051057e-01     NA     NA     FALSE     FALSE
#> 37  1.745946e+01 5.395852e-01  1e-06     NA     FALSE     FALSE
#> 38  2.073758e+01 8.177873e-01  1e-06     NA     FALSE     FALSE
#> 39  6.424033e+00 2.577174e+00  1e-06     NA     FALSE     FALSE
#> 40  4.804237e+01 2.871019e-01  1e-06     NA     FALSE     FALSE
#> 41  3.662241e+00 3.556134e+00  1e-06     NA     FALSE     FALSE
#> 42  6.686512e+00 1.687208e+00  1e-06     NA     FALSE     FALSE
#> 43  1.947520e+01 9.070029e-01  1e-06     NA     FALSE     FALSE
#> 44  1.239018e+01 4.046155e-01  1e-06     NA     FALSE     FALSE
#> 45  1.375468e+01 1.215325e+00  1e-06     NA     FALSE     FALSE
#> 46  3.279184e+01 5.316595e-01  1e-06     NA     FALSE     FALSE
#> 47  2.534045e+01 4.934418e-01     NA     NA     FALSE     FALSE
#> 48  8.986414e+01 2.223863e-01  1e-06     NA     FALSE     FALSE
#> 49 -6.689425e-02 3.138600e+02     NA     NA     FALSE     FALSE
#> 50 -8.436357e-02 1.958435e+02     NA     NA     FALSE     FALSE
#> 51  2.954513e-04 7.806432e+04  1e-06     NA      TRUE     FALSE
#> 52  2.964745e+01 4.366391e+00     NA     NA     FALSE     FALSE
#> 53  2.393592e+01 5.375130e+00     NA     NA     FALSE     FALSE
#> 54  1.416027e+01 9.541669e+00     NA     NA     FALSE     FALSE
#> 55  3.669880e+01 3.185114e+00     NA     NA     FALSE     FALSE
#> 56  8.467966e+00 1.120898e+01     NA     NA     FALSE     FALSE
#> 57  1.598080e+01 7.719867e+00     NA     NA     FALSE     FALSE
#> 58  1.389102e+01 5.660464e+00     NA     NA     FALSE     FALSE
#> 59  1.754485e+01 3.780916e+00     NA     NA     FALSE     FALSE
#> 60  2.276923e+01 6.552603e+00     NA     NA     FALSE     FALSE
#> 
#> --------------------------------------------------------------------------------

# let's get the groups:
get_groups(mg_model)
#> <list_of<
#>   tbl_df<
#>     id      : integer
#>     Gender  : factor<aa50b>
#>     grade   : integer
#>     agey    : integer
#>     agem    : integer
#>     school  : factor<b7a18>
#>     visual  : integer
#>     cubes   : integer
#>     paper   : integer
#>     flags   : integer
#>     general : integer
#>     paragrap: integer
#>     sentence: integer
#>     wordc   : integer
#>     wordm   : integer
#>     addition: integer
#>     code    : integer
#>     counting: integer
#>     straight: integer
#>     wordr   : integer
#>     numberr : integer
#>     figurer : integer
#>     object  : integer
#>     numberf : integer
#>     figurew : integer
#>     deduct  : integer
#>     numeric : integer
#>     problemr: integer
#>     series  : integer
#>     arithmet: integer
#>     paperrev: integer
#>     flagssub: integer
#>   >
#> >[2]>
#> $group_1
#> # A tibble: 145 × 32
#>       id Gender grade  agey  agem school      visual cubes paper flags general
#>    <int> <fct>  <int> <int> <int> <fct>        <int> <int> <int> <int>   <int>
#>  1   201 Male       7    13     0 Grant-White     23    19    13     4      46
#>  2   202 Female     7    11    10 Grant-White     33    22    12    17      43
#>  3   203 Male       7    12     6 Grant-White     34    24    14    22      36
#>  4   204 Male       7    11    11 Grant-White     29    23    12     9      38
#>  5   205 Male       7    12     5 Grant-White     16    25    11    10      51
#>  6   206 Female     7    12     6 Grant-White     30    25    12    20      42
#>  7   208 Female     7    12     8 Grant-White     36    33    19    36      69
#>  8   209 Female     7    11    11 Grant-White     28    25    10     9      35
#>  9   210 Female     7    12     5 Grant-White     30    25    15    11      32
#> 10   211 Female     7    12     5 Grant-White     20    25    13     6      39
#> # ℹ 135 more rows
#> # ℹ 21 more variables: paragrap <int>, sentence <int>, wordc <int>,
#> #   wordm <int>, addition <int>, code <int>, counting <int>, straight <int>,
#> #   wordr <int>, numberr <int>, figurer <int>, object <int>, numberf <int>,
#> #   figurew <int>, deduct <int>, numeric <int>, problemr <int>, series <int>,
#> #   arithmet <int>, paperrev <int>, flagssub <int>
#> 
#> $group_2
#> # A tibble: 156 × 32
#>       id Gender grade  agey  agem school  visual cubes paper flags general
#>    <int> <fct>  <int> <int> <int> <fct>    <int> <int> <int> <int>   <int>
#>  1     1 Male       7    13     1 Pasteur     20    31    12     3      40
#>  2     2 Female     7    13     7 Pasteur     32    21    12    17      34
#>  3     3 Female     7    13     1 Pasteur     27    21    12    15      20
#>  4     4 Male       7    13     2 Pasteur     32    31    16    24      42
#>  5     5 Female     7    12     2 Pasteur     29    19    12     7      37
#>  6     6 Female     7    14     1 Pasteur     32    20    11    18      31
#>  7     7 Male       7    12     1 Pasteur     17    24    12     8      40
#>  8     8 Female     7    12     2 Pasteur     34    25    13    15      29
#>  9     9 Female     7    13     0 Pasteur     27    23    11    12      29
#> 10    11 Female     7    12     5 Pasteur     21    21    10     6      33
#> # ℹ 146 more rows
#> # ℹ 21 more variables: paragrap <int>, sentence <int>, wordc <int>,
#> #   wordm <int>, addition <int>, code <int>, counting <int>, straight <int>,
#> #   wordr <int>, numberr <int>, figurer <int>, object <int>, numberf <int>,
#> #   figurew <int>, deduct <int>, numeric <int>, problemr <int>, series <int>,
#> #   arithmet <int>, paperrev <int>, flagssub <int>
#>