parameter_table_rcpp

creates a parameter table from a lavaan like syntax

Usage

parameter_table_rcpp(
  syntax,
  add_intercept,
  add_variance,
  add_exogenous_latent_covariances,
  add_exogenous_manifest_covariances,
  scale_latent_variance,
  scale_loading
)

Arguments

syntax

lavaan like syntax

add_intercept

should intercepts for manifest variables be automatically added?

add_variance

should variances for all variables be automatically added?

add_exogenous_latent_covariances

should covariances between exogenous latent variables be added automatically?

add_exogenous_manifest_covariances

should covariances between exogenous manifest variables be added automatically?

scale_latent_variance

should variances of latent variables be set to 1?

scale_loading

should the first loading of each latent variable be set to 1?

Value

parameter table