get_groups

returns a list of groups for a multi group model

Usage

get_groups(multi_group_model)

Arguments

multi_group_model

multi group model created with mxsem_group_by

Value

list with data for each group

Examples

# THE FOLLOWING EXAMPLE IS ADAPTED FROM
# https://openmx.ssri.psu.edu/docs/OpenMx/latest/_static/Rdoc/mxModel.html
library(mxsem)

model <- 'spatial =~ visual + cubes + paper
          verbal  =~ general + paragrap + sentence
          math    =~ numeric + series + arithmet'

mg_model <- mxsem(model = model,
                  data  = OpenMx::HS.ability.data) |>
  # we want separate models for all combinations of grades and schools:
  mxsem_group_by(grouping_variables = "school") |>
  mxTryHard()
#> Warning: This function is very experimental and may not yet work properly. Use with caution.
#> The following parameters will be the same across groups: 
#> The following parameters will be group specific: spatial→cubes, spatial→paper, verbal→paragrap, verbal→sentence, math→series, math→arithmet, visual↔visual, cubes↔cubes, paper↔paper, general↔general, paragrap↔paragrap, sentence↔sentence, numeric↔numeric, series↔series, arithmet↔arithmet, spatial↔spatial, spatial↔verbal, verbal↔verbal, spatial↔math, verbal↔math, math↔math, one→visual, one→cubes, one→paper, one→general, one→paragrap, one→sentence, one→numeric, one→series, one→arithmet
#> Running untitled1 with 60 parameters
#> 
#> Beginning initial fit attempt
#> Running untitled1 with 60 parameters
#> 
#>  Lowest minimum so far:  15969.1164364042
#>  Not all eigenvalues of the Hessian are positive: 402219341.891002, 24136125.6511393, 13629.6144673512, 13392.3232184624, 5863.79847475932, 4759.40902391292, 4360.27273268796, 1952.6181949062, 1825.43588373687, 1316.14320026041, 1149.69073403232, 357.020859626818, 139.587433865707, 71.9069260853621, 69.9404892014232, 46.269886249906, 42.6674235408757, 38.5891239225252, 24.7840310381383, 23.5787338184784, 22.843212497112, 21.5482242637395, 18.3282408584932, 15.9762169652866, 15.3060860161677, 13.9445602579177, 9.90204880539195, 6.10860281338663, 5.88811684225566, 5.50821258581185, 4.37657765716972, 4.36058933723348, 3.16509425753877, 2.38968859583729, 1.63544781213769, 1.58365930943974, 1.58124508589757, 1.38618515635273, 1.05930596548883, 0.739622567890229, 0.519989228275848, 0.479705953325449, 0.408746816359721, 0.367294348213997, 0.32782118531774, 0.18308465960063, 0.152239339155351, 0.0815597420870642, 0.0637842634380338, 0.0438074692131403, 0.0383380250844334, 0.0370122618603593, 0.0294527879632414, 0.0209545898869084, 0.0194967028842953, 0.00904037989428845, 0.00750007582366306, 0.00577013380997979, 2.1042732070824e-05, -0.000242048900633485
#> 
#> Beginning fit attempt 1 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#>  Not all eigenvalues of the Hessian are positive: 402219341.891002, 24136125.6511393, 13629.6144673512, 13392.3232184624, 5863.79847475932, 4759.40902391292, 4360.27273268796, 1952.6181949062, 1825.43588373687, 1316.14320026041, 1149.69073403232, 357.020859626818, 139.587433865707, 71.9069260853621, 69.9404892014232, 46.269886249906, 42.6674235408757, 38.5891239225252, 24.7840310381383, 23.5787338184784, 22.843212497112, 21.5482242637395, 18.3282408584932, 15.9762169652866, 15.3060860161677, 13.9445602579177, 9.90204880539195, 6.10860281338663, 5.88811684225566, 5.50821258581185, 4.37657765716972, 4.36058933723348, 3.16509425753877, 2.38968859583729, 1.63544781213769, 1.58365930943974, 1.58124508589757, 1.38618515635273, 1.05930596548883, 0.739622567890229, 0.519989228275848, 0.479705953325449, 0.408746816359721, 0.367294348213997, 0.32782118531774, 0.18308465960063, 0.152239339155351, 0.0815597420870642, 0.0637842634380338, 0.0438074692131403, 0.0383380250844334, 0.0370122618603593, 0.0294527879632414, 0.0209545898869084, 0.0194967028842953, 0.00904037989428845, 0.00750007582366306, 0.00577013380997979, 2.1042732070824e-05, -0.000242048900633485
#> 
#> Beginning fit attempt 2 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Lowest minimum so far:  15962.3267675686
#>  OpenMx status code 6 not in list of acceptable status codes, (0,0)
#>  Not all eigenvalues of the Hessian are positive: 1887234355.9862, 28722.2159335985, 8278.35720069326, 6146.87225990554, 4798.50604859787, 2141.90319713986, 1830.64662237646, 1092.22890525429, 340.743082538273, 142.234094637232, 71.4919416055596, 70.0951963747934, 46.3447876632895, 41.8993119861833, 40.4246535051776, 39.2331780472575, 25.0631252735779, 24.1616164453614, 22.9253808856384, 21.3602453895251, 17.1778140894903, 16.0338516500684, 15.4142079478497, 13.999469393913, 13.3111929572997, 11.2828673161444, 10.0534022268835, 6.7774317148991, 5.86981180995234, 5.62840558512586, 5.48160820744852, 4.43423070450347, 4.39486150524961, 3.15965107682328, 1.77637322351732, 1.70735179533731, 1.67147001752277, 1.49849235154138, 1.42386617147348, 1.10972818111665, 0.741245478078162, 0.49551134940068, 0.493261001344155, 0.481791239787085, 0.413105031891152, 0.336217125064198, 0.177523251137576, 0.172663414878747, 0.0796513591409654, 0.0750299214770689, 0.0408439841517805, 0.0375430292663172, 0.0339270541153296, 0.0322941605352615, 0.0215226992887029, 0.0104173161030819, 0.00549620777628884, 0.00294227090230414, 0.000228158502568061, -0.234144517075148
#> 
#> Beginning fit attempt 3 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Lowest minimum so far:  15915.879853312
#>  OpenMx status code 6 not in list of acceptable status codes, (0,0)
#> 
#> Beginning fit attempt 4 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Lowest minimum so far:  15915.8788735271
#> 
#> Solution found
#> 
#> 
#>  Solution found!  Final fit=15915.879 (started at 11279122)  (5 attempt(s): 5 valid, 0 errors)
#>  Start values from best fit:
#> 0.429550027894847,0.285948188558194,0.311937556833585,0.441910164186453,2.27753059517868,0.956902607349832,23.8789752641307,15.2709096805611,5.97877870997863,55.0898910719423,3.53292954109001,5.84840188296306,14.0506215671049,39.6450969246149,13.7102500690348,23.6058164742499,25.2239403139956,80.0621549486353,12.1132254663482,18.9979582451774,9.13478047653249,29.5792143865275,24.799908360986,14.3033923914879,44.847825396899,9.95163096634298,18.8481886648764,14.4550646656341,18.7515302079977,25.8275858788936,0.318362007059639,0.22236547701888,0.302325574824456,0.479133751792914,-476.303312534072,-137.11241527354,17.4616798034973,20.7353332299615,6.42415880340724,48.2269496741955,3.65949274493295,6.68652077828991,19.4578538771284,12.5504249249148,13.7489451897277,32.8036534745563,25.4005887000942,90.0598618489335,-0.0660648462579852,-0.0834171366253835,0.000287619939780028,29.6474367059424,23.9359646799896,14.1603089301026,36.6985580187814,8.46787821761521,15.9806649601491,13.891060728958,17.5449094843729,22.7691836133822

# let's summarize the results:
summarize_multi_group_model(mg_model)
#> 
#> 
#> --------------------------------------------------------------------------------
#> Common Parameters:
#>  [1] name      matrix    row       col       Estimate  Std.Error lbound   
#>  [8] ubound    lboundMet uboundMet
#> <0 rows> (or 0-length row.names)
#> --------------------------------------------------------------------------------
#> group_1:
#> # A tibble: 1 × 1
#>   school     
#>   <fct>      
#> 1 Grant-White
#> 
#>                         name              matrix      row      col   Estimate
#> 1      spatial→cubes_group_1 untitled1_group_1.A    cubes  spatial  0.4295500
#> 2      spatial→paper_group_1 untitled1_group_1.A    paper  spatial  0.2859482
#> 3    verbal→paragrap_group_1 untitled1_group_1.A paragrap   verbal  0.3119376
#> 4    verbal→sentence_group_1 untitled1_group_1.A sentence   verbal  0.4419102
#> 5        math→series_group_1 untitled1_group_1.A   series     math  2.2775306
#> 6      math→arithmet_group_1 untitled1_group_1.A arithmet     math  0.9569026
#> 7      visual↔visual_group_1 untitled1_group_1.S   visual   visual 23.8789753
#> 8        cubes↔cubes_group_1 untitled1_group_1.S    cubes    cubes 15.2709097
#> 9        paper↔paper_group_1 untitled1_group_1.S    paper    paper  5.9787787
#> 10   general↔general_group_1 untitled1_group_1.S  general  general 55.0898911
#> 11 paragrap↔paragrap_group_1 untitled1_group_1.S paragrap paragrap  3.5329295
#> 12 sentence↔sentence_group_1 untitled1_group_1.S sentence sentence  5.8484019
#> 13   numeric↔numeric_group_1 untitled1_group_1.S  numeric  numeric 14.0506216
#> 14     series↔series_group_1 untitled1_group_1.S   series   series 39.6450969
#> 15 arithmet↔arithmet_group_1 untitled1_group_1.S arithmet arithmet 13.7102501
#> 16   spatial↔spatial_group_1 untitled1_group_1.S  spatial  spatial 23.6058165
#> 17    spatial↔verbal_group_1 untitled1_group_1.S  spatial   verbal 25.2239403
#> 18     verbal↔verbal_group_1 untitled1_group_1.S   verbal   verbal 80.0621549
#> 19      spatial↔math_group_1 untitled1_group_1.S  spatial     math 12.1132255
#> 20       verbal↔math_group_1 untitled1_group_1.S   verbal     math 18.9979582
#> 21         math↔math_group_1 untitled1_group_1.S     math     math  9.1347805
#> 22        one→visual_group_1 untitled1_group_1.M        1   visual 29.5792144
#> 23         one→cubes_group_1 untitled1_group_1.M        1    cubes 24.7999084
#> 24         one→paper_group_1 untitled1_group_1.M        1    paper 14.3033924
#> 25       one→general_group_1 untitled1_group_1.M        1  general 44.8478254
#> 26      one→paragrap_group_1 untitled1_group_1.M        1 paragrap  9.9516310
#> 27      one→sentence_group_1 untitled1_group_1.M        1 sentence 18.8481887
#> 28       one→numeric_group_1 untitled1_group_1.M        1  numeric 14.4550647
#> 29        one→series_group_1 untitled1_group_1.M        1   series 18.7515302
#> 30      one→arithmet_group_1 untitled1_group_1.M        1 arithmet 25.8275859
#>     Std.Error lbound ubound lboundMet uboundMet
#> 1  24.9106347     NA     NA     FALSE     FALSE
#> 2  39.6763711     NA     NA     FALSE     FALSE
#> 3  93.5111076     NA     NA     FALSE     FALSE
#> 4  70.2303699     NA     NA     FALSE     FALSE
#> 5   9.3348704     NA     NA     FALSE     FALSE
#> 6  16.7129221     NA     NA     FALSE     FALSE
#> 7   0.5213751  1e-06     NA     FALSE     FALSE
#> 8   1.0285298  1e-06     NA     FALSE     FALSE
#> 9   2.5981227  1e-06     NA     FALSE     FALSE
#> 10  0.2526174  1e-06     NA     FALSE     FALSE
#> 11  3.4809127  1e-06     NA     FALSE     FALSE
#> 12  1.9309534  1e-06     NA     FALSE     FALSE
#> 13  1.0520995  1e-06     NA     FALSE     FALSE
#> 14  0.3252344  1e-06     NA     FALSE     FALSE
#> 15  1.0871809  1e-06     NA     FALSE     FALSE
#> 16  0.6512615  1e-06     NA     FALSE     FALSE
#> 17  0.6039066     NA     NA     FALSE     FALSE
#> 18  0.2716706  1e-06     NA     FALSE     FALSE
#> 19  1.8689215     NA     NA     FALSE     FALSE
#> 20  1.2106552     NA     NA     FALSE     FALSE
#> 21  2.2912573  1e-06     NA     FALSE     FALSE
#> 22  4.2135591     NA     NA     FALSE     FALSE
#> 23  5.9188150     NA     NA     FALSE     FALSE
#> 24  9.4067014     NA     NA     FALSE     FALSE
#> 25  2.9324943     NA     NA     FALSE     FALSE
#> 26 10.8874573     NA     NA     FALSE     FALSE
#> 27  8.1087804     NA     NA     FALSE     FALSE
#> 28  5.9850076     NA     NA     FALSE     FALSE
#> 29  3.3276329     NA     NA     FALSE     FALSE
#> 30  6.0872239     NA     NA     FALSE     FALSE
#> --------------------------------------------------------------------------------
#> group_2:
#> # A tibble: 1 × 1
#>   school 
#>   <fct>  
#> 1 Pasteur
#> 
#>                         name              matrix      row      col
#> 31     spatial→cubes_group_2 untitled1_group_2.A    cubes  spatial
#> 32     spatial→paper_group_2 untitled1_group_2.A    paper  spatial
#> 33   verbal→paragrap_group_2 untitled1_group_2.A paragrap   verbal
#> 34   verbal→sentence_group_2 untitled1_group_2.A sentence   verbal
#> 35       math→series_group_2 untitled1_group_2.A   series     math
#> 36     math→arithmet_group_2 untitled1_group_2.A arithmet     math
#> 37     visual↔visual_group_2 untitled1_group_2.S   visual   visual
#> 38       cubes↔cubes_group_2 untitled1_group_2.S    cubes    cubes
#> 39       paper↔paper_group_2 untitled1_group_2.S    paper    paper
#> 40   general↔general_group_2 untitled1_group_2.S  general  general
#> 41 paragrap↔paragrap_group_2 untitled1_group_2.S paragrap paragrap
#> 42 sentence↔sentence_group_2 untitled1_group_2.S sentence sentence
#> 43   numeric↔numeric_group_2 untitled1_group_2.S  numeric  numeric
#> 44     series↔series_group_2 untitled1_group_2.S   series   series
#> 45 arithmet↔arithmet_group_2 untitled1_group_2.S arithmet arithmet
#> 46   spatial↔spatial_group_2 untitled1_group_2.S  spatial  spatial
#> 47    spatial↔verbal_group_2 untitled1_group_2.S  spatial   verbal
#> 48     verbal↔verbal_group_2 untitled1_group_2.S   verbal   verbal
#> 49      spatial↔math_group_2 untitled1_group_2.S  spatial     math
#> 50       verbal↔math_group_2 untitled1_group_2.S   verbal     math
#> 51         math↔math_group_2 untitled1_group_2.S     math     math
#> 52        one→visual_group_2 untitled1_group_2.M        1   visual
#> 53         one→cubes_group_2 untitled1_group_2.M        1    cubes
#> 54         one→paper_group_2 untitled1_group_2.M        1    paper
#> 55       one→general_group_2 untitled1_group_2.M        1  general
#> 56      one→paragrap_group_2 untitled1_group_2.M        1 paragrap
#> 57      one→sentence_group_2 untitled1_group_2.M        1 sentence
#> 58       one→numeric_group_2 untitled1_group_2.M        1  numeric
#> 59        one→series_group_2 untitled1_group_2.M        1   series
#> 60      one→arithmet_group_2 untitled1_group_2.M        1 arithmet
#>         Estimate    Std.Error lbound ubound lboundMet uboundMet
#> 31  3.183620e-01 2.686028e+01     NA     NA     FALSE     FALSE
#> 32  2.223655e-01 4.787641e+01     NA     NA     FALSE     FALSE
#> 33  3.023256e-01 1.022126e+02     NA     NA     FALSE     FALSE
#> 34  4.791338e-01 7.130623e+01     NA     NA     FALSE     FALSE
#> 35 -4.763033e+02 7.027183e-02     NA     NA     FALSE     FALSE
#> 36 -1.371124e+02 1.036581e-01     NA     NA     FALSE     FALSE
#> 37  1.746168e+01 5.395075e-01  1e-06     NA     FALSE     FALSE
#> 38  2.073533e+01 8.181000e-01  1e-06     NA     FALSE     FALSE
#> 39  6.424159e+00 2.577086e+00  1e-06     NA     FALSE     FALSE
#> 40  4.822695e+01 2.853978e-01  1e-06     NA     FALSE     FALSE
#> 41  3.659493e+00 3.557228e+00  1e-06     NA     FALSE     FALSE
#> 42  6.686521e+00 1.686587e+00  1e-06     NA     FALSE     FALSE
#> 43  1.945785e+01 9.085896e-01  1e-06     NA     FALSE     FALSE
#> 44  1.255042e+01 4.036609e-01  1e-06     NA     FALSE     FALSE
#> 45  1.374895e+01 1.215431e+00  1e-06     NA     FALSE     FALSE
#> 46  3.280365e+01 5.314632e-01  1e-06     NA     FALSE     FALSE
#> 47  2.540059e+01 4.926481e-01     NA     NA     FALSE     FALSE
#> 48  9.005986e+01 2.216620e-01  1e-06     NA     FALSE     FALSE
#> 49 -6.606485e-02 3.175418e+02     NA     NA     FALSE     FALSE
#> 50 -8.341714e-02 1.978864e+02     NA     NA     FALSE     FALSE
#> 51  2.876199e-04 7.998125e+04  1e-06     NA      TRUE     FALSE
#> 52  2.964744e+01 4.365884e+00     NA     NA     FALSE     FALSE
#> 53  2.393596e+01 5.375414e+00     NA     NA     FALSE     FALSE
#> 54  1.416031e+01 9.541588e+00     NA     NA     FALSE     FALSE
#> 55  3.669856e+01 3.180128e+00     NA     NA     FALSE     FALSE
#> 56  8.467878e+00 1.120992e+01     NA     NA     FALSE     FALSE
#> 57  1.598066e+01 7.718639e+00     NA     NA     FALSE     FALSE
#> 58  1.389106e+01 5.662968e+00     NA     NA     FALSE     FALSE
#> 59  1.754491e+01 3.777278e+00     NA     NA     FALSE     FALSE
#> 60  2.276918e+01 6.552660e+00     NA     NA     FALSE     FALSE
#> 
#> --------------------------------------------------------------------------------

# let's get the groups:
get_groups(mg_model)
#> <list_of<
#>   tbl_df<
#>     id      : integer
#>     Gender  : factor<aa50b>
#>     grade   : integer
#>     agey    : integer
#>     agem    : integer
#>     school  : factor<b7a18>
#>     visual  : integer
#>     cubes   : integer
#>     paper   : integer
#>     flags   : integer
#>     general : integer
#>     paragrap: integer
#>     sentence: integer
#>     wordc   : integer
#>     wordm   : integer
#>     addition: integer
#>     code    : integer
#>     counting: integer
#>     straight: integer
#>     wordr   : integer
#>     numberr : integer
#>     figurer : integer
#>     object  : integer
#>     numberf : integer
#>     figurew : integer
#>     deduct  : integer
#>     numeric : integer
#>     problemr: integer
#>     series  : integer
#>     arithmet: integer
#>     paperrev: integer
#>     flagssub: integer
#>   >
#> >[2]>
#> $group_1
#> # A tibble: 145 × 32
#>       id Gender grade  agey  agem school      visual cubes paper flags general
#>    <int> <fct>  <int> <int> <int> <fct>        <int> <int> <int> <int>   <int>
#>  1   201 Male       7    13     0 Grant-White     23    19    13     4      46
#>  2   202 Female     7    11    10 Grant-White     33    22    12    17      43
#>  3   203 Male       7    12     6 Grant-White     34    24    14    22      36
#>  4   204 Male       7    11    11 Grant-White     29    23    12     9      38
#>  5   205 Male       7    12     5 Grant-White     16    25    11    10      51
#>  6   206 Female     7    12     6 Grant-White     30    25    12    20      42
#>  7   208 Female     7    12     8 Grant-White     36    33    19    36      69
#>  8   209 Female     7    11    11 Grant-White     28    25    10     9      35
#>  9   210 Female     7    12     5 Grant-White     30    25    15    11      32
#> 10   211 Female     7    12     5 Grant-White     20    25    13     6      39
#> # ℹ 135 more rows
#> # ℹ 21 more variables: paragrap <int>, sentence <int>, wordc <int>,
#> #   wordm <int>, addition <int>, code <int>, counting <int>, straight <int>,
#> #   wordr <int>, numberr <int>, figurer <int>, object <int>, numberf <int>,
#> #   figurew <int>, deduct <int>, numeric <int>, problemr <int>, series <int>,
#> #   arithmet <int>, paperrev <int>, flagssub <int>
#> 
#> $group_2
#> # A tibble: 156 × 32
#>       id Gender grade  agey  agem school  visual cubes paper flags general
#>    <int> <fct>  <int> <int> <int> <fct>    <int> <int> <int> <int>   <int>
#>  1     1 Male       7    13     1 Pasteur     20    31    12     3      40
#>  2     2 Female     7    13     7 Pasteur     32    21    12    17      34
#>  3     3 Female     7    13     1 Pasteur     27    21    12    15      20
#>  4     4 Male       7    13     2 Pasteur     32    31    16    24      42
#>  5     5 Female     7    12     2 Pasteur     29    19    12     7      37
#>  6     6 Female     7    14     1 Pasteur     32    20    11    18      31
#>  7     7 Male       7    12     1 Pasteur     17    24    12     8      40
#>  8     8 Female     7    12     2 Pasteur     34    25    13    15      29
#>  9     9 Female     7    13     0 Pasteur     27    23    11    12      29
#> 10    11 Female     7    12     5 Pasteur     21    21    10     6      33
#> # ℹ 146 more rows
#> # ℹ 21 more variables: paragrap <int>, sentence <int>, wordc <int>,
#> #   wordm <int>, addition <int>, code <int>, counting <int>, straight <int>,
#> #   wordr <int>, numberr <int>, figurer <int>, object <int>, numberf <int>,
#> #   figurew <int>, deduct <int>, numeric <int>, problemr <int>, series <int>,
#> #   arithmet <int>, paperrev <int>, flagssub <int>
#>