get_groups

returns a list of groups for a multi group model

Usage

get_groups(multi_group_model)

Arguments

multi_group_model

multi group model created with mxsem_group_by

Value

list with data for each group

Examples

# THE FOLLOWING EXAMPLE IS ADAPTED FROM
# https://openmx.ssri.psu.edu/docs/OpenMx/latest/_static/Rdoc/mxModel.html
library(mxsem)

model <- 'spatial =~ visual + cubes + paper
          verbal  =~ general + paragrap + sentence
          math    =~ numeric + series + arithmet'

mg_model <- mxsem(model = model,
                  data  = OpenMx::HS.ability.data) |>
  # we want separate models for all combinations of grades and schools:
  mxsem_group_by(grouping_variables = "school") |>
  mxTryHard()
#> Warning: This function is very experimental and may not yet work properly. Use with caution.
#> The following parameters will be the same across groups: 
#> The following parameters will be group specific: spatial→cubes, spatial→paper, verbal→paragrap, verbal→sentence, math→series, math→arithmet, visual↔visual, cubes↔cubes, paper↔paper, general↔general, paragrap↔paragrap, sentence↔sentence, numeric↔numeric, series↔series, arithmet↔arithmet, spatial↔spatial, spatial↔verbal, verbal↔verbal, spatial↔math, verbal↔math, math↔math, one→visual, one→cubes, one→paper, one→general, one→paragrap, one→sentence, one→numeric, one→series, one→arithmet
#> Running untitled1 with 60 parameters
#> 
#> Beginning initial fit attempt
#> Running untitled1 with 60 parameters
#> 
#>  Lowest minimum so far:  15969.1164364042
#>  Not all eigenvalues of the Hessian are positive: 402219341.891002, 24136125.6511393, 13629.6144673512, 13392.3232184624, 5863.79847475932, 4759.40902391292, 4360.27273268796, 1952.6181949062, 1825.43588373687, 1316.14320026041, 1149.69073403232, 357.020859626818, 139.587433865707, 71.9069260853621, 69.9404892014232, 46.269886249906, 42.6674235408757, 38.5891239225252, 24.7840310381383, 23.5787338184784, 22.843212497112, 21.5482242637395, 18.3282408584932, 15.9762169652866, 15.3060860161677, 13.9445602579177, 9.90204880539195, 6.10860281338663, 5.88811684225566, 5.50821258581185, 4.37657765716972, 4.36058933723348, 3.16509425753877, 2.38968859583729, 1.63544781213769, 1.58365930943974, 1.58124508589757, 1.38618515635273, 1.05930596548883, 0.739622567890229, 0.519989228275848, 0.479705953325449, 0.408746816359721, 0.367294348213997, 0.32782118531774, 0.18308465960063, 0.152239339155351, 0.0815597420870642, 0.0637842634380338, 0.0438074692131403, 0.0383380250844334, 0.0370122618603593, 0.0294527879632414, 0.0209545898869084, 0.0194967028842953, 0.00904037989428845, 0.00750007582366306, 0.00577013380997979, 2.1042732070824e-05, -0.000242048900633485
#> 
#> Beginning fit attempt 1 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#>  Not all eigenvalues of the Hessian are positive: 402219341.891002, 24136125.6511393, 13629.6144673512, 13392.3232184624, 5863.79847475932, 4759.40902391292, 4360.27273268796, 1952.6181949062, 1825.43588373687, 1316.14320026041, 1149.69073403232, 357.020859626818, 139.587433865707, 71.9069260853621, 69.9404892014232, 46.269886249906, 42.6674235408757, 38.5891239225252, 24.7840310381383, 23.5787338184784, 22.843212497112, 21.5482242637395, 18.3282408584932, 15.9762169652866, 15.3060860161677, 13.9445602579177, 9.90204880539195, 6.10860281338663, 5.88811684225566, 5.50821258581185, 4.37657765716972, 4.36058933723348, 3.16509425753877, 2.38968859583729, 1.63544781213769, 1.58365930943974, 1.58124508589757, 1.38618515635273, 1.05930596548883, 0.739622567890229, 0.519989228275848, 0.479705953325449, 0.408746816359721, 0.367294348213997, 0.32782118531774, 0.18308465960063, 0.152239339155351, 0.0815597420870642, 0.0637842634380338, 0.0438074692131403, 0.0383380250844334, 0.0370122618603593, 0.0294527879632414, 0.0209545898869084, 0.0194967028842953, 0.00904037989428845, 0.00750007582366306, 0.00577013380997979, 2.1042732070824e-05, -0.000242048900633485
#> 
#> Beginning fit attempt 2 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Fit attempt generated errors
#> 
#> Beginning fit attempt 3 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Fit attempt generated errors
#> 
#> Beginning fit attempt 4 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Fit attempt generated errors
#> 
#> Beginning fit attempt 5 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Fit attempt generated errors
#> 
#> Beginning fit attempt 6 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Fit attempt generated errors
#> 
#> Beginning fit attempt 7 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Fit attempt worse than current best:  17151.6951254755 vs 15969.1164364042
#> 
#> Beginning fit attempt 8 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Fit attempt generated errors
#> 
#> Beginning fit attempt 9 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Fit attempt generated errors
#> 
#> Beginning fit attempt 10 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Fit attempt generated errors
#> 
#> Retry limit reached
#> 
#> 
#>  Retry limit reached; Best fit=15969.116 (started at 11279122)  (11 attempt(s): 3 valid, 8 errors)
#>  Start values from best fit:
#> 60.8143531025139,39.3962175259006,0.317998814580115,0.451827404187577,2.20260520599226,0.989106583746482,47.4738715344667,15.8045744522389,6.39850850662049,56.2242197929114,3.54169332012277,5.76487193516061,13.8981607565243,42.4906085870776,13.0746285334224,0.000996339548332554,0.18159003993062,75.0231275985784,0.0869324564168227,18.166016481142,9.04101453124251,29.5795185585207,24.7999512725151,14.3034410274732,44.8484478505408,9.95165703577297,18.8481198532474,14.4550723085255,18.75168896941,25.827500965918,0.316510689590047,0.221301241755882,0.300906490281909,0.47688706513643,-283.253114172045,-82.2836414942679,17.3497712235414,20.7554438523889,6.43067730090438,47.927274940026,3.66433304439884,6.70017466010673,19.5292656575528,12.9167298967888,13.7214635216479,33.1728376592433,25.924826595274,91.5247689194335,-0.111899584613787,-0.142127473273197,0.000810452094491224,29.6476092573724,23.9359476855704,14.160298281103,36.6988333333838,8.46800179450177,15.980882916537,13.8910397904594,17.5450152161267,22.7692757272316

# let's summarize the results:
summarize_multi_group_model(mg_model)
#> 
#> 
#> --------------------------------------------------------------------------------
#> Common Parameters:
#>  [1] name      matrix    row       col       Estimate  Std.Error lbound   
#>  [8] ubound    lboundMet uboundMet
#> <0 rows> (or 0-length row.names)
#> --------------------------------------------------------------------------------
#> group_1:
#> # A tibble: 1 × 1
#>   school     
#>   <fct>      
#> 1 Grant-White
#> 
#>                         name              matrix      row      col     Estimate
#> 1      spatial→cubes_group_1 untitled1_group_1.A    cubes  spatial 6.081435e+01
#> 2      spatial→paper_group_1 untitled1_group_1.A    paper  spatial 3.939622e+01
#> 3    verbal→paragrap_group_1 untitled1_group_1.A paragrap   verbal 3.179988e-01
#> 4    verbal→sentence_group_1 untitled1_group_1.A sentence   verbal 4.518274e-01
#> 5        math→series_group_1 untitled1_group_1.A   series     math 2.202605e+00
#> 6      math→arithmet_group_1 untitled1_group_1.A arithmet     math 9.891066e-01
#> 7      visual↔visual_group_1 untitled1_group_1.S   visual   visual 4.747387e+01
#> 8        cubes↔cubes_group_1 untitled1_group_1.S    cubes    cubes 1.580457e+01
#> 9        paper↔paper_group_1 untitled1_group_1.S    paper    paper 6.398509e+00
#> 10   general↔general_group_1 untitled1_group_1.S  general  general 5.622422e+01
#> 11 paragrap↔paragrap_group_1 untitled1_group_1.S paragrap paragrap 3.541693e+00
#> 12 sentence↔sentence_group_1 untitled1_group_1.S sentence sentence 5.764872e+00
#> 13   numeric↔numeric_group_1 untitled1_group_1.S  numeric  numeric 1.389816e+01
#> 14     series↔series_group_1 untitled1_group_1.S   series   series 4.249061e+01
#> 15 arithmet↔arithmet_group_1 untitled1_group_1.S arithmet arithmet 1.307463e+01
#> 16   spatial↔spatial_group_1 untitled1_group_1.S  spatial  spatial 9.963395e-04
#> 17    spatial↔verbal_group_1 untitled1_group_1.S  spatial   verbal 1.815900e-01
#> 18     verbal↔verbal_group_1 untitled1_group_1.S   verbal   verbal 7.502313e+01
#> 19      spatial↔math_group_1 untitled1_group_1.S  spatial     math 8.693246e-02
#> 20       verbal↔math_group_1 untitled1_group_1.S   verbal     math 1.816602e+01
#> 21         math↔math_group_1 untitled1_group_1.S     math     math 9.041015e+00
#> 22        one→visual_group_1 untitled1_group_1.M        1   visual 2.957952e+01
#> 23         one→cubes_group_1 untitled1_group_1.M        1    cubes 2.479995e+01
#> 24         one→paper_group_1 untitled1_group_1.M        1    paper 1.430344e+01
#> 25       one→general_group_1 untitled1_group_1.M        1  general 4.484845e+01
#> 26      one→paragrap_group_1 untitled1_group_1.M        1 paragrap 9.951657e+00
#> 27      one→sentence_group_1 untitled1_group_1.M        1 sentence 1.884812e+01
#> 28       one→numeric_group_1 untitled1_group_1.M        1  numeric 1.445507e+01
#> 29        one→series_group_1 untitled1_group_1.M        1   series 1.875169e+01
#> 30      one→arithmet_group_1 untitled1_group_1.M        1 arithmet 2.582750e+01
#>      Std.Error lbound ubound lboundMet uboundMet
#> 1           NA     NA     NA     FALSE     FALSE
#> 2           NA     NA     NA     FALSE     FALSE
#> 3   0.03321365     NA     NA     FALSE     FALSE
#> 4   0.04597164     NA     NA     FALSE     FALSE
#> 5   0.36418842     NA     NA     FALSE     FALSE
#> 6   0.17185720     NA     NA     FALSE     FALSE
#> 7   5.59373703  1e-06     NA     FALSE     FALSE
#> 8   2.35935228  1e-06     NA     FALSE     FALSE
#> 9   0.96628706  1e-06     NA     FALSE     FALSE
#> 10  8.61027005  1e-06     NA     FALSE     FALSE
#> 11  0.63493902  1e-06     NA     FALSE     FALSE
#> 12  1.17541854  1e-06     NA     FALSE     FALSE
#> 13  2.07896104  1e-06     NA     FALSE     FALSE
#> 14  7.33019332  1e-06     NA     FALSE     FALSE
#> 15  1.98808788  1e-06     NA     FALSE     FALSE
#> 16          NA  1e-06     NA      TRUE     FALSE
#> 17          NA     NA     NA     FALSE     FALSE
#> 18 14.45835712  1e-06     NA     FALSE     FALSE
#> 19          NA     NA     NA     FALSE     FALSE
#> 20  3.77608019     NA     NA     FALSE     FALSE
#> 21  2.47328018  1e-06     NA     FALSE     FALSE
#> 22  0.57220064     NA     NA     FALSE     FALSE
#> 23  0.36661916     NA     NA     FALSE     FALSE
#> 24  0.23407777     NA     NA     FALSE     FALSE
#> 25  0.95139650     NA     NA     FALSE     FALSE
#> 26  0.27703211     NA     NA     FALSE     FALSE
#> 27  0.38129308     NA     NA     FALSE     FALSE
#> 28  0.39774486     NA     NA     FALSE     FALSE
#> 29  0.77171071     NA     NA     FALSE     FALSE
#> 30  0.38880614     NA     NA     FALSE     FALSE
#> --------------------------------------------------------------------------------
#> group_2:
#> # A tibble: 1 × 1
#>   school 
#>   <fct>  
#> 1 Pasteur
#> 
#>                         name              matrix      row      col
#> 31     spatial→cubes_group_2 untitled1_group_2.A    cubes  spatial
#> 32     spatial→paper_group_2 untitled1_group_2.A    paper  spatial
#> 33   verbal→paragrap_group_2 untitled1_group_2.A paragrap   verbal
#> 34   verbal→sentence_group_2 untitled1_group_2.A sentence   verbal
#> 35       math→series_group_2 untitled1_group_2.A   series     math
#> 36     math→arithmet_group_2 untitled1_group_2.A arithmet     math
#> 37     visual↔visual_group_2 untitled1_group_2.S   visual   visual
#> 38       cubes↔cubes_group_2 untitled1_group_2.S    cubes    cubes
#> 39       paper↔paper_group_2 untitled1_group_2.S    paper    paper
#> 40   general↔general_group_2 untitled1_group_2.S  general  general
#> 41 paragrap↔paragrap_group_2 untitled1_group_2.S paragrap paragrap
#> 42 sentence↔sentence_group_2 untitled1_group_2.S sentence sentence
#> 43   numeric↔numeric_group_2 untitled1_group_2.S  numeric  numeric
#> 44     series↔series_group_2 untitled1_group_2.S   series   series
#> 45 arithmet↔arithmet_group_2 untitled1_group_2.S arithmet arithmet
#> 46   spatial↔spatial_group_2 untitled1_group_2.S  spatial  spatial
#> 47    spatial↔verbal_group_2 untitled1_group_2.S  spatial   verbal
#> 48     verbal↔verbal_group_2 untitled1_group_2.S   verbal   verbal
#> 49      spatial↔math_group_2 untitled1_group_2.S  spatial     math
#> 50       verbal↔math_group_2 untitled1_group_2.S   verbal     math
#> 51         math↔math_group_2 untitled1_group_2.S     math     math
#> 52        one→visual_group_2 untitled1_group_2.M        1   visual
#> 53         one→cubes_group_2 untitled1_group_2.M        1    cubes
#> 54         one→paper_group_2 untitled1_group_2.M        1    paper
#> 55       one→general_group_2 untitled1_group_2.M        1  general
#> 56      one→paragrap_group_2 untitled1_group_2.M        1 paragrap
#> 57      one→sentence_group_2 untitled1_group_2.M        1 sentence
#> 58       one→numeric_group_2 untitled1_group_2.M        1  numeric
#> 59        one→series_group_2 untitled1_group_2.M        1   series
#> 60      one→arithmet_group_2 untitled1_group_2.M        1 arithmet
#>         Estimate    Std.Error lbound ubound lboundMet uboundMet
#> 31  3.165107e-01 9.177452e-02     NA     NA     FALSE     FALSE
#> 32  2.213012e-01 5.274598e-02     NA     NA     FALSE     FALSE
#> 33  3.009065e-01 2.776029e-02     NA     NA     FALSE     FALSE
#> 34  4.768871e-01 4.033572e-02     NA     NA     FALSE     FALSE
#> 35 -2.832531e+02 2.961343e+02     NA     NA     FALSE     FALSE
#> 36 -8.228364e+01 8.726269e+01     NA     NA     FALSE     FALSE
#> 37  1.734977e+01 6.224204e+00  1e-06     NA     FALSE     FALSE
#> 38  2.075544e+01 2.526952e+00  1e-06     NA     FALSE     FALSE
#> 39  6.430677e+00 8.088100e-01  1e-06     NA     FALSE     FALSE
#> 40  4.792727e+01 7.427063e+00  1e-06     NA     FALSE     FALSE
#> 41  3.664333e+00 6.401276e-01  1e-06     NA     FALSE     FALSE
#> 42  6.700175e+00 1.371850e+00  1e-06     NA     FALSE     FALSE
#> 43  1.952927e+01 2.215541e+00  1e-06     NA     FALSE     FALSE
#> 44  1.291673e+01 1.106943e+01  1e-06     NA     FALSE     FALSE
#> 45  1.372146e+01 1.804431e+00  1e-06     NA     FALSE     FALSE
#> 46  3.317284e+01 8.013851e+00  1e-06     NA     FALSE     FALSE
#> 47  2.592483e+01 6.532703e+00     NA     NA     FALSE     FALSE
#> 48  9.152477e+01 1.584652e+01  1e-06     NA     FALSE     FALSE
#> 49 -1.118996e-01 1.186531e-01     NA     NA     FALSE     FALSE
#> 50 -1.421275e-01 1.513987e-01     NA     NA     FALSE     FALSE
#> 51  8.104521e-04 1.697619e-03  1e-06     NA      TRUE     FALSE
#> 52  2.964761e+01 5.690910e-01     NA     NA     FALSE     FALSE
#> 53  2.393595e+01 3.928750e-01     NA     NA     FALSE     FALSE
#> 54  1.416030e+01 2.272368e-01     NA     NA     FALSE     FALSE
#> 55  3.669883e+01 9.454795e-01     NA     NA     FALSE     FALSE
#> 56  8.468002e+00 2.767894e-01     NA     NA     FALSE     FALSE
#> 57  1.598088e+01 4.199758e-01     NA     NA     FALSE     FALSE
#> 58  1.389104e+01 3.538246e-01     NA     NA     FALSE     FALSE
#> 59  1.754502e+01 7.068417e-01     NA     NA     FALSE     FALSE
#> 60  2.276928e+01 3.509032e-01     NA     NA     FALSE     FALSE
#> 
#> --------------------------------------------------------------------------------

# let's get the groups:
get_groups(mg_model)
#> <list_of<
#>   tbl_df<
#>     id      : integer
#>     Gender  : factor<aa50b>
#>     grade   : integer
#>     agey    : integer
#>     agem    : integer
#>     school  : factor<b7a18>
#>     visual  : integer
#>     cubes   : integer
#>     paper   : integer
#>     flags   : integer
#>     general : integer
#>     paragrap: integer
#>     sentence: integer
#>     wordc   : integer
#>     wordm   : integer
#>     addition: integer
#>     code    : integer
#>     counting: integer
#>     straight: integer
#>     wordr   : integer
#>     numberr : integer
#>     figurer : integer
#>     object  : integer
#>     numberf : integer
#>     figurew : integer
#>     deduct  : integer
#>     numeric : integer
#>     problemr: integer
#>     series  : integer
#>     arithmet: integer
#>     paperrev: integer
#>     flagssub: integer
#>   >
#> >[2]>
#> $group_1
#> # A tibble: 145 × 32
#>       id Gender grade  agey  agem school      visual cubes paper flags general
#>    <int> <fct>  <int> <int> <int> <fct>        <int> <int> <int> <int>   <int>
#>  1   201 Male       7    13     0 Grant-White     23    19    13     4      46
#>  2   202 Female     7    11    10 Grant-White     33    22    12    17      43
#>  3   203 Male       7    12     6 Grant-White     34    24    14    22      36
#>  4   204 Male       7    11    11 Grant-White     29    23    12     9      38
#>  5   205 Male       7    12     5 Grant-White     16    25    11    10      51
#>  6   206 Female     7    12     6 Grant-White     30    25    12    20      42
#>  7   208 Female     7    12     8 Grant-White     36    33    19    36      69
#>  8   209 Female     7    11    11 Grant-White     28    25    10     9      35
#>  9   210 Female     7    12     5 Grant-White     30    25    15    11      32
#> 10   211 Female     7    12     5 Grant-White     20    25    13     6      39
#> # ℹ 135 more rows
#> # ℹ 21 more variables: paragrap <int>, sentence <int>, wordc <int>,
#> #   wordm <int>, addition <int>, code <int>, counting <int>, straight <int>,
#> #   wordr <int>, numberr <int>, figurer <int>, object <int>, numberf <int>,
#> #   figurew <int>, deduct <int>, numeric <int>, problemr <int>, series <int>,
#> #   arithmet <int>, paperrev <int>, flagssub <int>
#> 
#> $group_2
#> # A tibble: 156 × 32
#>       id Gender grade  agey  agem school  visual cubes paper flags general
#>    <int> <fct>  <int> <int> <int> <fct>    <int> <int> <int> <int>   <int>
#>  1     1 Male       7    13     1 Pasteur     20    31    12     3      40
#>  2     2 Female     7    13     7 Pasteur     32    21    12    17      34
#>  3     3 Female     7    13     1 Pasteur     27    21    12    15      20
#>  4     4 Male       7    13     2 Pasteur     32    31    16    24      42
#>  5     5 Female     7    12     2 Pasteur     29    19    12     7      37
#>  6     6 Female     7    14     1 Pasteur     32    20    11    18      31
#>  7     7 Male       7    12     1 Pasteur     17    24    12     8      40
#>  8     8 Female     7    12     2 Pasteur     34    25    13    15      29
#>  9     9 Female     7    13     0 Pasteur     27    23    11    12      29
#> 10    11 Female     7    12     5 Pasteur     21    21    10     6      33
#> # ℹ 146 more rows
#> # ℹ 21 more variables: paragrap <int>, sentence <int>, wordc <int>,
#> #   wordm <int>, addition <int>, code <int>, counting <int>, straight <int>,
#> #   wordr <int>, numberr <int>, figurer <int>, object <int>, numberf <int>,
#> #   figurew <int>, deduct <int>, numeric <int>, problemr <int>, series <int>,
#> #   arithmet <int>, paperrev <int>, flagssub <int>
#>