get_groups

returns a list of groups for a multi group model

Usage

get_groups(multi_group_model)

Arguments

multi_group_model

multi group model created with mxsem_group_by

Value

list with data for each group

Examples

# THE FOLLOWING EXAMPLE IS ADAPTED FROM
# https://openmx.ssri.psu.edu/docs/OpenMx/latest/_static/Rdoc/mxModel.html
library(mxsem)

model <- 'spatial =~ visual + cubes + paper
          verbal  =~ general + paragrap + sentence
          math    =~ numeric + series + arithmet'

mg_model <- mxsem(model = model,
                  data  = OpenMx::HS.ability.data) |>
  # we want separate models for all combinations of grades and schools:
  mxsem_group_by(grouping_variables = "school") |>
  mxTryHard()
#> Warning: This function is very experimental and may not yet work properly. Use with caution.
#> The following parameters will be the same across groups: 
#> The following parameters will be group specific: spatial→cubes, spatial→paper, verbal→paragrap, verbal→sentence, math→series, math→arithmet, visual↔visual, cubes↔cubes, paper↔paper, general↔general, paragrap↔paragrap, sentence↔sentence, numeric↔numeric, series↔series, arithmet↔arithmet, spatial↔spatial, spatial↔verbal, verbal↔verbal, spatial↔math, verbal↔math, math↔math, one→visual, one→cubes, one→paper, one→general, one→paragrap, one→sentence, one→numeric, one→series, one→arithmet
#> Running untitled1 with 60 parameters
#> 
#> Beginning initial fit attempt
#> Running untitled1 with 60 parameters
#> 
#>  Lowest minimum so far:  15915.8948628543
#>  OpenMx status code 6 not in list of acceptable status codes, (0,0)
#> 
#> Beginning fit attempt 1 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Lowest minimum so far:  15915.8919985549
#>  Not all eigenvalues of the Hessian are positive: 2881324677.33359, 36009.3516286094, 12023.6703681719, 5729.48946656132, 4937.78018906414, 1953.20743294049, 1890.59548458872, 1141.93768709214, 787.053832922359, 354.299869949574, 310.059988684748, 140.308070837012, 71.960524679851, 69.7737543984885, 46.2892558097408, 44.9970707911168, 41.9999858871418, 24.9093324641401, 23.5191899890895, 22.8577228952339, 20.9223694392101, 18.3025744933685, 16.0851747896718, 16.018845743146, 15.310385828649, 9.90808200029694, 8.45369857439435, 5.88962751326757, 5.53488330138983, 4.35940260488893, 4.18490800861019, 3.49096374055486, 3.16683866275163, 2.82736036010931, 1.59910088820855, 1.51503506990849, 1.43956890190422, 1.15853458220828, 1.05434253829188, 0.740020523034631, 0.518708350463338, 0.502847939424892, 0.492307464631392, 0.411213793573148, 0.327917491475641, 0.240774010995978, 0.184333452645261, 0.150707405584368, 0.0854720418961566, 0.0836640606223451, 0.0414555910630677, 0.0385806931620717, 0.0370749105068254, 0.0339971038622095, 0.0295537215170684, 0.021067635452236, 0.00785326267622024, 0.00757395103844831, 0.00250913156290403, -1.55655451604865e-06
#> 
#> Beginning fit attempt 2 of at maximum 10 extra tries
#> Running untitled1 with 60 parameters
#> 
#>  Lowest minimum so far:  15915.8919074347
#> 
#> Solution found
#> 
#> 
#>  Solution found!  Final fit=15915.892 (started at 11279122)  (3 attempt(s): 3 valid, 0 errors)
#>  Start values from best fit:
#> 0.429696324732396,0.286024397585417,0.312259740235881,0.442363859837346,2.27691771976176,0.956883492242165,23.8894216200827,15.2682075393357,5.97877715105563,55.1488164616464,3.53184435633436,5.84673040324623,14.0527084575859,39.6645663103217,13.6972693242364,23.5809085727407,25.1619866544482,79.8223569879943,12.1068191452977,18.9621175559387,9.13346418813602,29.5793258731495,24.7999951477868,14.3034555329686,44.8483120475232,9.9517398564783,18.8482871851452,14.4551793966412,18.751732741285,25.8276029806873,0.318407659498922,0.222407964517981,0.302368730073987,0.479270268445379,-464.258471510988,-133.869634378721,17.4577160519638,20.7378606419435,6.4239542394658,48.039286873544,3.66215052845238,6.6871834935806,19.4772292706673,12.4486942624279,13.7506093409822,32.7899172753595,25.3475083356719,89.8901007982516,-0.0678139344851871,-0.0855209149857523,0.00030325245185672,29.6474204666946,23.9358983573264,14.1602534807989,36.6986412140479,8.46793193567707,15.9807408169199,13.8910283708663,17.5448422432478,22.7692116297089

# let's summarize the results:
summarize_multi_group_model(mg_model)
#> 
#> 
#> --------------------------------------------------------------------------------
#> Common Parameters:
#>  [1] name      matrix    row       col       Estimate  Std.Error lbound   
#>  [8] ubound    lboundMet uboundMet
#> <0 rows> (or 0-length row.names)
#> --------------------------------------------------------------------------------
#> group_1:
#> # A tibble: 1 × 1
#>   school     
#>   <fct>      
#> 1 Grant-White
#> 
#>                         name              matrix      row      col   Estimate
#> 1      spatial→cubes_group_1 untitled1_group_1.A    cubes  spatial  0.4296963
#> 2      spatial→paper_group_1 untitled1_group_1.A    paper  spatial  0.2860244
#> 3    verbal→paragrap_group_1 untitled1_group_1.A paragrap   verbal  0.3122597
#> 4    verbal→sentence_group_1 untitled1_group_1.A sentence   verbal  0.4423639
#> 5        math→series_group_1 untitled1_group_1.A   series     math  2.2769177
#> 6      math→arithmet_group_1 untitled1_group_1.A arithmet     math  0.9568835
#> 7      visual↔visual_group_1 untitled1_group_1.S   visual   visual 23.8894216
#> 8        cubes↔cubes_group_1 untitled1_group_1.S    cubes    cubes 15.2682075
#> 9        paper↔paper_group_1 untitled1_group_1.S    paper    paper  5.9787772
#> 10   general↔general_group_1 untitled1_group_1.S  general  general 55.1488165
#> 11 paragrap↔paragrap_group_1 untitled1_group_1.S paragrap paragrap  3.5318444
#> 12 sentence↔sentence_group_1 untitled1_group_1.S sentence sentence  5.8467304
#> 13   numeric↔numeric_group_1 untitled1_group_1.S  numeric  numeric 14.0527085
#> 14     series↔series_group_1 untitled1_group_1.S   series   series 39.6645663
#> 15 arithmet↔arithmet_group_1 untitled1_group_1.S arithmet arithmet 13.6972693
#> 16   spatial↔spatial_group_1 untitled1_group_1.S  spatial  spatial 23.5809086
#> 17    spatial↔verbal_group_1 untitled1_group_1.S  spatial   verbal 25.1619867
#> 18     verbal↔verbal_group_1 untitled1_group_1.S   verbal   verbal 79.8223570
#> 19      spatial↔math_group_1 untitled1_group_1.S  spatial     math 12.1068191
#> 20       verbal↔math_group_1 untitled1_group_1.S   verbal     math 18.9621176
#> 21         math↔math_group_1 untitled1_group_1.S     math     math  9.1334642
#> 22        one→visual_group_1 untitled1_group_1.M        1   visual 29.5793259
#> 23         one→cubes_group_1 untitled1_group_1.M        1    cubes 24.7999951
#> 24         one→paper_group_1 untitled1_group_1.M        1    paper 14.3034555
#> 25       one→general_group_1 untitled1_group_1.M        1  general 44.8483120
#> 26      one→paragrap_group_1 untitled1_group_1.M        1 paragrap  9.9517399
#> 27      one→sentence_group_1 untitled1_group_1.M        1 sentence 18.8482872
#> 28       one→numeric_group_1 untitled1_group_1.M        1  numeric 14.4551794
#> 29        one→series_group_1 untitled1_group_1.M        1   series 18.7517327
#> 30      one→arithmet_group_1 untitled1_group_1.M        1 arithmet 25.8276030
#>     Std.Error lbound ubound lboundMet uboundMet
#> 1  24.9037882     NA     NA     FALSE     FALSE
#> 2  39.6622211     NA     NA     FALSE     FALSE
#> 3  93.4253200     NA     NA     FALSE     FALSE
#> 4  70.1639596     NA     NA     FALSE     FALSE
#> 5   9.3352733     NA     NA     FALSE     FALSE
#> 6  16.7201928     NA     NA     FALSE     FALSE
#> 7   0.5212566  1e-06     NA     FALSE     FALSE
#> 8   1.0288039  1e-06     NA     FALSE     FALSE
#> 9   2.5980818  1e-06     NA     FALSE     FALSE
#> 10  0.2523914  1e-06     NA     FALSE     FALSE
#> 11  3.4804341  1e-06     NA     FALSE     FALSE
#> 12  1.9305246  1e-06     NA     FALSE     FALSE
#> 13  1.0515824  1e-06     NA     FALSE     FALSE
#> 14  0.3250416  1e-06     NA     FALSE     FALSE
#> 15  1.0888323  1e-06     NA     FALSE     FALSE
#> 16  0.6516031  1e-06     NA     FALSE     FALSE
#> 17  0.6049615     NA     NA     FALSE     FALSE
#> 18  0.2724801  1e-06     NA     FALSE     FALSE
#> 19  1.8694247     NA     NA     FALSE     FALSE
#> 20  1.2120947     NA     NA     FALSE     FALSE
#> 21  2.2905729  1e-06     NA     FALSE     FALSE
#> 22  4.2133898     NA     NA     FALSE     FALSE
#> 23  5.9192583     NA     NA     FALSE     FALSE
#> 24  9.4067036     NA     NA     FALSE     FALSE
#> 25  2.9318542     NA     NA     FALSE     FALSE
#> 26 10.8878455     NA     NA     FALSE     FALSE
#> 27  8.1087416     NA     NA     FALSE     FALSE
#> 28  5.9844863     NA     NA     FALSE     FALSE
#> 29  3.3271878     NA     NA     FALSE     FALSE
#> 30  6.0895770     NA     NA     FALSE     FALSE
#> --------------------------------------------------------------------------------
#> group_2:
#> # A tibble: 1 × 1
#>   school 
#>   <fct>  
#> 1 Pasteur
#> 
#>                         name              matrix      row      col
#> 31     spatial→cubes_group_2 untitled1_group_2.A    cubes  spatial
#> 32     spatial→paper_group_2 untitled1_group_2.A    paper  spatial
#> 33   verbal→paragrap_group_2 untitled1_group_2.A paragrap   verbal
#> 34   verbal→sentence_group_2 untitled1_group_2.A sentence   verbal
#> 35       math→series_group_2 untitled1_group_2.A   series     math
#> 36     math→arithmet_group_2 untitled1_group_2.A arithmet     math
#> 37     visual↔visual_group_2 untitled1_group_2.S   visual   visual
#> 38       cubes↔cubes_group_2 untitled1_group_2.S    cubes    cubes
#> 39       paper↔paper_group_2 untitled1_group_2.S    paper    paper
#> 40   general↔general_group_2 untitled1_group_2.S  general  general
#> 41 paragrap↔paragrap_group_2 untitled1_group_2.S paragrap paragrap
#> 42 sentence↔sentence_group_2 untitled1_group_2.S sentence sentence
#> 43   numeric↔numeric_group_2 untitled1_group_2.S  numeric  numeric
#> 44     series↔series_group_2 untitled1_group_2.S   series   series
#> 45 arithmet↔arithmet_group_2 untitled1_group_2.S arithmet arithmet
#> 46   spatial↔spatial_group_2 untitled1_group_2.S  spatial  spatial
#> 47    spatial↔verbal_group_2 untitled1_group_2.S  spatial   verbal
#> 48     verbal↔verbal_group_2 untitled1_group_2.S   verbal   verbal
#> 49      spatial↔math_group_2 untitled1_group_2.S  spatial     math
#> 50       verbal↔math_group_2 untitled1_group_2.S   verbal     math
#> 51         math↔math_group_2 untitled1_group_2.S     math     math
#> 52        one→visual_group_2 untitled1_group_2.M        1   visual
#> 53         one→cubes_group_2 untitled1_group_2.M        1    cubes
#> 54         one→paper_group_2 untitled1_group_2.M        1    paper
#> 55       one→general_group_2 untitled1_group_2.M        1  general
#> 56      one→paragrap_group_2 untitled1_group_2.M        1 paragrap
#> 57      one→sentence_group_2 untitled1_group_2.M        1 sentence
#> 58       one→numeric_group_2 untitled1_group_2.M        1  numeric
#> 59        one→series_group_2 untitled1_group_2.M        1   series
#> 60      one→arithmet_group_2 untitled1_group_2.M        1 arithmet
#>         Estimate    Std.Error lbound ubound lboundMet uboundMet
#> 31  3.184077e-01 2.685600e+01     NA     NA     FALSE     FALSE
#> 32  2.224080e-01 4.787092e+01     NA     NA     FALSE     FALSE
#> 33  3.023687e-01 1.021691e+02     NA     NA     FALSE     FALSE
#> 34  4.792703e-01 7.129261e+01     NA     NA     FALSE     FALSE
#> 35 -4.642585e+02 7.225536e-02     NA     NA     FALSE     FALSE
#> 36 -1.338696e+02 1.064656e-01     NA     NA     FALSE     FALSE
#> 37  1.745772e+01 5.399522e-01  1e-06     NA     FALSE     FALSE
#> 38  2.073786e+01 8.178606e-01  1e-06     NA     FALSE     FALSE
#> 39  6.423954e+00 2.577448e+00  1e-06     NA     FALSE     FALSE
#> 40  4.803929e+01 2.871769e-01  1e-06     NA     FALSE     FALSE
#> 41  3.662151e+00 3.556516e+00  1e-06     NA     FALSE     FALSE
#> 42  6.687183e+00 1.687191e+00  1e-06     NA     FALSE     FALSE
#> 43  1.947723e+01 9.069229e-01  1e-06     NA     FALSE     FALSE
#> 44  1.244869e+01 4.049034e-01  1e-06     NA     FALSE     FALSE
#> 45  1.375061e+01 1.215510e+00  1e-06     NA     FALSE     FALSE
#> 46  3.278992e+01 5.320073e-01  1e-06     NA     FALSE     FALSE
#> 47  2.534751e+01 4.935328e-01     NA     NA     FALSE     FALSE
#> 48  8.989010e+01 2.223162e-01  1e-06     NA     FALSE     FALSE
#> 49 -6.781393e-02 3.098339e+02     NA     NA     FALSE     FALSE
#> 50 -8.552091e-02 1.932178e+02     NA     NA     FALSE     FALSE
#> 51  3.032525e-04 7.601277e+04  1e-06     NA      TRUE     FALSE
#> 52  2.964742e+01 4.366879e+00     NA     NA     FALSE     FALSE
#> 53  2.393590e+01 5.375123e+00     NA     NA     FALSE     FALSE
#> 54  1.416025e+01 9.541766e+00     NA     NA     FALSE     FALSE
#> 55  3.669864e+01 3.185122e+00     NA     NA     FALSE     FALSE
#> 56  8.467932e+00 1.120904e+01     NA     NA     FALSE     FALSE
#> 57  1.598074e+01 7.719790e+00     NA     NA     FALSE     FALSE
#> 58  1.389103e+01 5.660159e+00     NA     NA     FALSE     FALSE
#> 59  1.754484e+01 3.780316e+00     NA     NA     FALSE     FALSE
#> 60  2.276921e+01 6.552695e+00     NA     NA     FALSE     FALSE
#> 
#> --------------------------------------------------------------------------------

# let's get the groups:
get_groups(mg_model)
#> <list_of<
#>   tbl_df<
#>     id      : integer
#>     Gender  : factor<aa50b>
#>     grade   : integer
#>     agey    : integer
#>     agem    : integer
#>     school  : factor<b7a18>
#>     visual  : integer
#>     cubes   : integer
#>     paper   : integer
#>     flags   : integer
#>     general : integer
#>     paragrap: integer
#>     sentence: integer
#>     wordc   : integer
#>     wordm   : integer
#>     addition: integer
#>     code    : integer
#>     counting: integer
#>     straight: integer
#>     wordr   : integer
#>     numberr : integer
#>     figurer : integer
#>     object  : integer
#>     numberf : integer
#>     figurew : integer
#>     deduct  : integer
#>     numeric : integer
#>     problemr: integer
#>     series  : integer
#>     arithmet: integer
#>     paperrev: integer
#>     flagssub: integer
#>   >
#> >[2]>
#> $group_1
#> # A tibble: 145 × 32
#>       id Gender grade  agey  agem school      visual cubes paper flags general
#>    <int> <fct>  <int> <int> <int> <fct>        <int> <int> <int> <int>   <int>
#>  1   201 Male       7    13     0 Grant-White     23    19    13     4      46
#>  2   202 Female     7    11    10 Grant-White     33    22    12    17      43
#>  3   203 Male       7    12     6 Grant-White     34    24    14    22      36
#>  4   204 Male       7    11    11 Grant-White     29    23    12     9      38
#>  5   205 Male       7    12     5 Grant-White     16    25    11    10      51
#>  6   206 Female     7    12     6 Grant-White     30    25    12    20      42
#>  7   208 Female     7    12     8 Grant-White     36    33    19    36      69
#>  8   209 Female     7    11    11 Grant-White     28    25    10     9      35
#>  9   210 Female     7    12     5 Grant-White     30    25    15    11      32
#> 10   211 Female     7    12     5 Grant-White     20    25    13     6      39
#> # ℹ 135 more rows
#> # ℹ 21 more variables: paragrap <int>, sentence <int>, wordc <int>,
#> #   wordm <int>, addition <int>, code <int>, counting <int>, straight <int>,
#> #   wordr <int>, numberr <int>, figurer <int>, object <int>, numberf <int>,
#> #   figurew <int>, deduct <int>, numeric <int>, problemr <int>, series <int>,
#> #   arithmet <int>, paperrev <int>, flagssub <int>
#> 
#> $group_2
#> # A tibble: 156 × 32
#>       id Gender grade  agey  agem school  visual cubes paper flags general
#>    <int> <fct>  <int> <int> <int> <fct>    <int> <int> <int> <int>   <int>
#>  1     1 Male       7    13     1 Pasteur     20    31    12     3      40
#>  2     2 Female     7    13     7 Pasteur     32    21    12    17      34
#>  3     3 Female     7    13     1 Pasteur     27    21    12    15      20
#>  4     4 Male       7    13     2 Pasteur     32    31    16    24      42
#>  5     5 Female     7    12     2 Pasteur     29    19    12     7      37
#>  6     6 Female     7    14     1 Pasteur     32    20    11    18      31
#>  7     7 Male       7    12     1 Pasteur     17    24    12     8      40
#>  8     8 Female     7    12     2 Pasteur     34    25    13    15      29
#>  9     9 Female     7    13     0 Pasteur     27    23    11    12      29
#> 10    11 Female     7    12     5 Pasteur     21    21    10     6      33
#> # ℹ 146 more rows
#> # ℹ 21 more variables: paragrap <int>, sentence <int>, wordc <int>,
#> #   wordm <int>, addition <int>, code <int>, counting <int>, straight <int>,
#> #   wordr <int>, numberr <int>, figurer <int>, object <int>, numberf <int>,
#> #   figurew <int>, deduct <int>, numeric <int>, problemr <int>, series <int>,
#> #   arithmet <int>, paperrev <int>, flagssub <int>
#>