mxsem_group_by

creates a multi-group model from an OpenMx model.

Usage

mxsem_group_by(
  mxModel,
  grouping_variables,
  parameters = c(".*"),
  use_grepl = TRUE
)

Arguments

mxModel

mxModel with the entire data

grouping_variables

Variables used to split the data in groups

parameters

the parameters that should be group specific. By default all parameters are group specific.

use_grepl

if set to TRUE, grepl is used to check which parameters are group specific. For instance, if parameters = "a" and use_grepl = TRUE, all parameters whose label contains the letter "a" will be group specific. If use_grep = FALSE only the parameter that has the label "a" is group specific.

Value

mxModel with multiple groups. Use get_groups to extract the groups

Details

mxsem_group_by creates a multi-group model by splitting the data found in an mxModel object using dplyr's group_by function. The general idea is as follows:

1. The function extracts the data from mxModel 2. The data is split using the group_by function of dplyr with the variables in grouping_variables 3. a separate model is set up for each group. All parameters that match those specified in the parameters argument are group specific

**Warning**: The multi-group model may differ from **lavaan**! For instance, **lavaan** will automatically set the latent variances for all but the first group free if the loadings are fixed to equality. Such automatic procedures are not yet implemented in **mxsem**.

Examples

# THE FOLLOWING EXAMPLE IS ADAPTED FROM
# https://openmx.ssri.psu.edu/docs/OpenMx/latest/_static/Rdoc/mxModel.html
library(mxsem)

model <- 'spatial =~ visual + cubes + paper
          verbal  =~ general + paragrap + sentence
          math    =~ numeric + series + arithmet'

mg_model <- mxsem(model = model,
                  data  = OpenMx::HS.ability.data) |>
  # we want separate models for all combinations of grades and schools:
  mxsem_group_by(grouping_variables = "school") |>
  mxTryHard()
#> Warning: This function is very experimental and may not yet work properly. Use with caution.
#> The following parameters will be the same across groups: 
#> The following parameters will be group specific: spatial→cubes, spatial→paper, verbal→paragrap, verbal→sentence, math→series, math→arithmet, visual↔visual, cubes↔cubes, paper↔paper, general↔general, paragrap↔paragrap, sentence↔sentence, numeric↔numeric, series↔series, arithmet↔arithmet, spatial↔spatial, spatial↔verbal, verbal↔verbal, spatial↔math, verbal↔math, math↔math, one→visual, one→cubes, one→paper, one→general, one→paragrap, one→sentence, one→numeric, one→series, one→arithmet
#> Running untitled27 with 60 parameters
#> 
#> Beginning initial fit attempt
#> Running untitled27 with 60 parameters
#> 
#>  Lowest minimum so far:  15969.1164364042
#>  Not all eigenvalues of the Hessian are positive: 402219341.891002, 24136125.6511393, 13629.6144673512, 13392.3232184624, 5863.79847475932, 4759.40902391292, 4360.27273268796, 1952.6181949062, 1825.43588373687, 1316.14320026041, 1149.69073403232, 357.020859626818, 139.587433865707, 71.9069260853621, 69.9404892014232, 46.269886249906, 42.6674235408757, 38.5891239225252, 24.7840310381383, 23.5787338184784, 22.843212497112, 21.5482242637395, 18.3282408584932, 15.9762169652866, 15.3060860161677, 13.9445602579177, 9.90204880539195, 6.10860281338663, 5.88811684225566, 5.50821258581185, 4.37657765716972, 4.36058933723348, 3.16509425753877, 2.38968859583729, 1.63544781213769, 1.58365930943974, 1.58124508589757, 1.38618515635273, 1.05930596548883, 0.739622567890229, 0.519989228275848, 0.479705953325449, 0.408746816359721, 0.367294348213997, 0.32782118531774, 0.18308465960063, 0.152239339155351, 0.0815597420870642, 0.0637842634380338, 0.0438074692131403, 0.0383380250844334, 0.0370122618603593, 0.0294527879632414, 0.0209545898869084, 0.0194967028842953, 0.00904037989428845, 0.00750007582366306, 0.00577013380997979, 2.1042732070824e-05, -0.000242048900633485
#> 
#> Beginning fit attempt 1 of at maximum 10 extra tries
#> Running untitled27 with 60 parameters
#>  Not all eigenvalues of the Hessian are positive: 402219341.891002, 24136125.6511393, 13629.6144673512, 13392.3232184624, 5863.79847475932, 4759.40902391292, 4360.27273268796, 1952.6181949062, 1825.43588373687, 1316.14320026041, 1149.69073403232, 357.020859626818, 139.587433865707, 71.9069260853621, 69.9404892014232, 46.269886249906, 42.6674235408757, 38.5891239225252, 24.7840310381383, 23.5787338184784, 22.843212497112, 21.5482242637395, 18.3282408584932, 15.9762169652866, 15.3060860161677, 13.9445602579177, 9.90204880539195, 6.10860281338663, 5.88811684225566, 5.50821258581185, 4.37657765716972, 4.36058933723348, 3.16509425753877, 2.38968859583729, 1.63544781213769, 1.58365930943974, 1.58124508589757, 1.38618515635273, 1.05930596548883, 0.739622567890229, 0.519989228275848, 0.479705953325449, 0.408746816359721, 0.367294348213997, 0.32782118531774, 0.18308465960063, 0.152239339155351, 0.0815597420870642, 0.0637842634380338, 0.0438074692131403, 0.0383380250844334, 0.0370122618603593, 0.0294527879632414, 0.0209545898869084, 0.0194967028842953, 0.00904037989428845, 0.00750007582366306, 0.00577013380997979, 2.1042732070824e-05, -0.000242048900633485
#> 
#> Beginning fit attempt 2 of at maximum 10 extra tries
#> Running untitled27 with 60 parameters
#> 
#>  Lowest minimum so far:  15964.2485650762
#>  OpenMx status code 6 not in list of acceptable status codes, (0,0)
#>  Not all eigenvalues of the Hessian are positive: 2267708511.59165, 31537.3506596481, 8309.5667066467, 6201.1049475517, 4783.53419571009, 2190.21612284954, 1829.08994929642, 1163.79781102052, 745.524477338961, 360.735826642838, 157.848628403877, 70.9348875680571, 69.8435205179456, 66.6563018300491, 46.2637132908214, 44.7651249840484, 42.4943345966734, 24.8038752869437, 24.6275560241145, 22.8842917105404, 21.5761059876698, 18.4111342718383, 15.9823052382046, 15.2973440914761, 13.8166695661515, 9.83624692750449, 9.21241547690291, 6.4367808222943, 5.82753723721818, 5.50110112070068, 4.33303094935915, 4.31283369982507, 4.09894862072929, 3.1679690505364, 2.26002672278524, 1.59374455639454, 1.50902572015324, 1.43936186859176, 1.24781547692186, 1.13826182635679, 0.741017495838948, 0.524840762330478, 0.47497601063778, 0.407566303288832, 0.40527437764235, 0.326226215648104, 0.18371574154488, 0.145425752556015, 0.0701928838363377, 0.0695690318352953, 0.0414707250502998, 0.0383698290934375, 0.0375428516351267, 0.0304083014700978, 0.0190726179563928, 0.00941106556191836, 0.00494904400037415, 0.00271428300933824, -0.000117477063269266, -0.121542478675577
#> 
#> Beginning fit attempt 3 of at maximum 10 extra tries
#> Running untitled27 with 60 parameters
#> 
#>  Lowest minimum so far:  15921.1772815874
#>  OpenMx status code 6 not in list of acceptable status codes, (0,0)
#> 
#> Beginning fit attempt 4 of at maximum 10 extra tries
#> Running untitled27 with 60 parameters
#> 
#>  Lowest minimum so far:  15915.8761182841
#>  Not all eigenvalues of the Hessian are positive: 3226647476.34444, 38101.7828502995, 12713.851166443, 5731.19958678547, 4937.71631050429, 1953.73153491299, 1890.57748040273, 1142.04644089855, 787.03945980779, 354.305920252278, 310.043215332078, 140.245837743417, 71.9591256740992, 69.7730193385638, 46.2898576959964, 44.996957734005, 41.9986799294326, 24.9089399828436, 23.521043810156, 22.8589756227791, 20.922195371164, 18.3025089163733, 16.0852814573973, 16.0204423017274, 15.3105349951744, 9.90815259760566, 8.45362098013221, 5.88930061847693, 5.53227677178077, 4.35913266484781, 4.18481371834664, 3.48999103838485, 3.16698666521557, 2.82725070600076, 1.59895068125415, 1.5143749289344, 1.43975978538953, 1.15811499955601, 1.05456354292287, 0.739852274730976, 0.518553372102194, 0.502828040994879, 0.492106396689765, 0.411311275313168, 0.327925834529546, 0.240656938753255, 0.184238194707978, 0.150581237014464, 0.0854828726956325, 0.083644501866098, 0.0414359261371014, 0.0385668806091524, 0.0368235230559041, 0.0339196833978119, 0.0295559308850907, 0.0210310654640062, 0.00784908997895948, 0.0075696151695046, 0.0023816805846059, -1.68915201541812e-06
#> 
#> Beginning fit attempt 5 of at maximum 10 extra tries
#> Running untitled27 with 60 parameters
#> 
#>  Fit attempt generated errors
#> 
#> Beginning fit attempt 6 of at maximum 10 extra tries
#> Running untitled27 with 60 parameters
#> 
#>  Fit attempt worse than current best:  16281.8571519725 vs 15915.8761182841
#> 
#> Beginning fit attempt 7 of at maximum 10 extra tries
#> Running untitled27 with 60 parameters
#> 
#>  Fit attempt worse than current best:  16231.7688620527 vs 15915.8761182841
#> 
#> Beginning fit attempt 8 of at maximum 10 extra tries
#> Running untitled27 with 60 parameters
#> 
#>  Fit attempt generated errors
#> 
#> Beginning fit attempt 9 of at maximum 10 extra tries
#> Running untitled27 with 60 parameters
#> 
#>  Fit attempt worse than current best:  16107.0463588753 vs 15915.8761182841
#> 
#> Beginning fit attempt 10 of at maximum 10 extra tries
#> Running untitled27 with 60 parameters
#> 
#>  Fit attempt generated errors
#> 
#> Retry limit reached
#> 
#> 
#>  Retry limit reached; Best fit=15915.876 (started at 11279122)  (11 attempt(s): 8 valid, 3 errors)
#>  Start values from best fit:
#> 0.429696215641264,0.28602257726335,0.312259256900375,0.442363218659684,2.27697411313191,0.956891385252506,23.8899253685641,15.2681640705744,5.97877312972873,55.1490898740784,3.53185032661942,5.84674925133987,14.0528103694985,39.6624773713467,13.6972563826441,23.5808174218677,25.1603015529069,79.8214556708296,12.1064105055324,18.9611272321113,9.1330389330197,29.5793095663011,24.8000027820073,14.3034480836489,44.8482735108994,9.95172470177203,18.8482766994658,14.4551732005758,18.7517282668398,25.8275842228506,0.318415459842259,0.222407917380397,0.302360900560574,0.479261681999414,-477.625080417658,-137.57492254795,17.4579883999208,20.7375580301921,6.42385050622528,48.0333332323942,3.66226483018403,6.68718420853211,19.4750935187873,12.3893358510448,13.7545397069528,32.7977513420364,25.3564850545507,89.8997951056086,-0.0659441363407981,-0.0831573908384755,0.000286987129454263,29.6474342096626,23.9358949006147,14.1602535719142,36.6987097186207,8.46794653224228,15.9807655614192,13.8910239991387,17.5448613814941,22.7692265399749

# let's summarize the results:
summarize_multi_group_model(mg_model)
#> 
#> 
#> --------------------------------------------------------------------------------
#> Common Parameters:
#>  [1] name      matrix    row       col       Estimate  Std.Error lbound   
#>  [8] ubound    lboundMet uboundMet
#> <0 rows> (or 0-length row.names)
#> --------------------------------------------------------------------------------
#> group_1:
#> # A tibble: 1 × 1
#>   school     
#>   <fct>      
#> 1 Grant-White
#> 
#>                         name               matrix      row      col   Estimate
#> 1      spatial→cubes_group_1 untitled27_group_1.A    cubes  spatial  0.4296962
#> 2      spatial→paper_group_1 untitled27_group_1.A    paper  spatial  0.2860226
#> 3    verbal→paragrap_group_1 untitled27_group_1.A paragrap   verbal  0.3122593
#> 4    verbal→sentence_group_1 untitled27_group_1.A sentence   verbal  0.4423632
#> 5        math→series_group_1 untitled27_group_1.A   series     math  2.2769741
#> 6      math→arithmet_group_1 untitled27_group_1.A arithmet     math  0.9568914
#> 7      visual↔visual_group_1 untitled27_group_1.S   visual   visual 23.8899254
#> 8        cubes↔cubes_group_1 untitled27_group_1.S    cubes    cubes 15.2681641
#> 9        paper↔paper_group_1 untitled27_group_1.S    paper    paper  5.9787731
#> 10   general↔general_group_1 untitled27_group_1.S  general  general 55.1490899
#> 11 paragrap↔paragrap_group_1 untitled27_group_1.S paragrap paragrap  3.5318503
#> 12 sentence↔sentence_group_1 untitled27_group_1.S sentence sentence  5.8467493
#> 13   numeric↔numeric_group_1 untitled27_group_1.S  numeric  numeric 14.0528104
#> 14     series↔series_group_1 untitled27_group_1.S   series   series 39.6624774
#> 15 arithmet↔arithmet_group_1 untitled27_group_1.S arithmet arithmet 13.6972564
#> 16   spatial↔spatial_group_1 untitled27_group_1.S  spatial  spatial 23.5808174
#> 17    spatial↔verbal_group_1 untitled27_group_1.S  spatial   verbal 25.1603016
#> 18     verbal↔verbal_group_1 untitled27_group_1.S   verbal   verbal 79.8214557
#> 19      spatial↔math_group_1 untitled27_group_1.S  spatial     math 12.1064105
#> 20       verbal↔math_group_1 untitled27_group_1.S   verbal     math 18.9611272
#> 21         math↔math_group_1 untitled27_group_1.S     math     math  9.1330389
#> 22        one→visual_group_1 untitled27_group_1.M        1   visual 29.5793096
#> 23         one→cubes_group_1 untitled27_group_1.M        1    cubes 24.8000028
#> 24         one→paper_group_1 untitled27_group_1.M        1    paper 14.3034481
#> 25       one→general_group_1 untitled27_group_1.M        1  general 44.8482735
#> 26      one→paragrap_group_1 untitled27_group_1.M        1 paragrap  9.9517247
#> 27      one→sentence_group_1 untitled27_group_1.M        1 sentence 18.8482767
#> 28       one→numeric_group_1 untitled27_group_1.M        1  numeric 14.4551732
#> 29        one→series_group_1 untitled27_group_1.M        1   series 18.7517283
#> 30      one→arithmet_group_1 untitled27_group_1.M        1 arithmet 25.8275842
#>     Std.Error lbound ubound lboundMet uboundMet
#> 1  24.9031145     NA     NA     FALSE     FALSE
#> 2  39.6621874     NA     NA     FALSE     FALSE
#> 3  93.4249747     NA     NA     FALSE     FALSE
#> 4  70.1635186     NA     NA     FALSE     FALSE
#> 5   9.3346165     NA     NA     FALSE     FALSE
#> 6  16.7183407     NA     NA     FALSE     FALSE
#> 7   0.5211978  1e-06     NA     FALSE     FALSE
#> 8   1.0287423  1e-06     NA     FALSE     FALSE
#> 9   2.5979022  1e-06     NA     FALSE     FALSE
#> 10  0.2523516  1e-06     NA     FALSE     FALSE
#> 11  3.4804273  1e-06     NA     FALSE     FALSE
#> 12  1.9303047  1e-06     NA     FALSE     FALSE
#> 13  1.0514939  1e-06     NA     FALSE     FALSE
#> 14  0.3250340  1e-06     NA     FALSE     FALSE
#> 15  1.0888335  1e-06     NA     FALSE     FALSE
#> 16  0.6515309  1e-06     NA     FALSE     FALSE
#> 17  0.6049041     NA     NA     FALSE     FALSE
#> 18  0.2724705  1e-06     NA     FALSE     FALSE
#> 19  1.8692656     NA     NA     FALSE     FALSE
#> 20  1.2121168     NA     NA     FALSE     FALSE
#> 21  2.2906053  1e-06     NA     FALSE     FALSE
#> 22  4.2133325     NA     NA     FALSE     FALSE
#> 23  5.9192558     NA     NA     FALSE     FALSE
#> 24  9.4066863     NA     NA     FALSE     FALSE
#> 25  2.9318416     NA     NA     FALSE     FALSE
#> 26 10.8878009     NA     NA     FALSE     FALSE
#> 27  8.1087111     NA     NA     FALSE     FALSE
#> 28  5.9844786     NA     NA     FALSE     FALSE
#> 29  3.3272408     NA     NA     FALSE     FALSE
#> 30  6.0895872     NA     NA     FALSE     FALSE
#> --------------------------------------------------------------------------------
#> group_2:
#> # A tibble: 1 × 1
#>   school 
#>   <fct>  
#> 1 Pasteur
#> 
#>                         name               matrix      row      col
#> 31     spatial→cubes_group_2 untitled27_group_2.A    cubes  spatial
#> 32     spatial→paper_group_2 untitled27_group_2.A    paper  spatial
#> 33   verbal→paragrap_group_2 untitled27_group_2.A paragrap   verbal
#> 34   verbal→sentence_group_2 untitled27_group_2.A sentence   verbal
#> 35       math→series_group_2 untitled27_group_2.A   series     math
#> 36     math→arithmet_group_2 untitled27_group_2.A arithmet     math
#> 37     visual↔visual_group_2 untitled27_group_2.S   visual   visual
#> 38       cubes↔cubes_group_2 untitled27_group_2.S    cubes    cubes
#> 39       paper↔paper_group_2 untitled27_group_2.S    paper    paper
#> 40   general↔general_group_2 untitled27_group_2.S  general  general
#> 41 paragrap↔paragrap_group_2 untitled27_group_2.S paragrap paragrap
#> 42 sentence↔sentence_group_2 untitled27_group_2.S sentence sentence
#> 43   numeric↔numeric_group_2 untitled27_group_2.S  numeric  numeric
#> 44     series↔series_group_2 untitled27_group_2.S   series   series
#> 45 arithmet↔arithmet_group_2 untitled27_group_2.S arithmet arithmet
#> 46   spatial↔spatial_group_2 untitled27_group_2.S  spatial  spatial
#> 47    spatial↔verbal_group_2 untitled27_group_2.S  spatial   verbal
#> 48     verbal↔verbal_group_2 untitled27_group_2.S   verbal   verbal
#> 49      spatial↔math_group_2 untitled27_group_2.S  spatial     math
#> 50       verbal↔math_group_2 untitled27_group_2.S   verbal     math
#> 51         math↔math_group_2 untitled27_group_2.S     math     math
#> 52        one→visual_group_2 untitled27_group_2.M        1   visual
#> 53         one→cubes_group_2 untitled27_group_2.M        1    cubes
#> 54         one→paper_group_2 untitled27_group_2.M        1    paper
#> 55       one→general_group_2 untitled27_group_2.M        1  general
#> 56      one→paragrap_group_2 untitled27_group_2.M        1 paragrap
#> 57      one→sentence_group_2 untitled27_group_2.M        1 sentence
#> 58       one→numeric_group_2 untitled27_group_2.M        1  numeric
#> 59        one→series_group_2 untitled27_group_2.M        1   series
#> 60      one→arithmet_group_2 untitled27_group_2.M        1 arithmet
#>         Estimate    Std.Error lbound ubound lboundMet uboundMet
#> 31  3.184155e-01 2.685581e+01     NA     NA     FALSE     FALSE
#> 32  2.224079e-01 4.787111e+01     NA     NA     FALSE     FALSE
#> 33  3.023609e-01 1.021700e+02     NA     NA     FALSE     FALSE
#> 34  4.792617e-01 7.129402e+01     NA     NA     FALSE     FALSE
#> 35 -4.776251e+02 7.029282e-02     NA     NA     FALSE     FALSE
#> 36 -1.375749e+02 1.035666e-01     NA     NA     FALSE     FALSE
#> 37  1.745799e+01 5.397869e-01  1e-06     NA     FALSE     FALSE
#> 38  2.073756e+01 8.178362e-01  1e-06     NA     FALSE     FALSE
#> 39  6.423851e+00 2.577210e+00  1e-06     NA     FALSE     FALSE
#> 40  4.803333e+01 2.872107e-01  1e-06     NA     FALSE     FALSE
#> 41  3.662265e+00 3.556409e+00  1e-06     NA     FALSE     FALSE
#> 42  6.687184e+00 1.687004e+00  1e-06     NA     FALSE     FALSE
#> 43  1.947509e+01 9.069856e-01  1e-06     NA     FALSE     FALSE
#> 44  1.238934e+01 4.043737e-01  1e-06     NA     FALSE     FALSE
#> 45  1.375454e+01 1.215371e+00  1e-06     NA     FALSE     FALSE
#> 46  3.279775e+01 5.318101e-01  1e-06     NA     FALSE     FALSE
#> 47  2.535649e+01 4.933745e-01     NA     NA     FALSE     FALSE
#> 48  8.989980e+01 2.222953e-01  1e-06     NA     FALSE     FALSE
#> 49 -6.594414e-02 3.184536e+02     NA     NA     FALSE     FALSE
#> 50 -8.315739e-02 1.986120e+02     NA     NA     FALSE     FALSE
#> 51  2.869871e-04 8.033216e+04  1e-06     NA      TRUE     FALSE
#> 52  2.964743e+01 4.366655e+00     NA     NA     FALSE     FALSE
#> 53  2.393589e+01 5.375138e+00     NA     NA     FALSE     FALSE
#> 54  1.416025e+01 9.541786e+00     NA     NA     FALSE     FALSE
#> 55  3.669871e+01 3.185244e+00     NA     NA     FALSE     FALSE
#> 56  8.467947e+00 1.120898e+01     NA     NA     FALSE     FALSE
#> 57  1.598077e+01 7.719812e+00     NA     NA     FALSE     FALSE
#> 58  1.389102e+01 5.660456e+00     NA     NA     FALSE     FALSE
#> 59  1.754486e+01 3.779648e+00     NA     NA     FALSE     FALSE
#> 60  2.276923e+01 6.552711e+00     NA     NA     FALSE     FALSE
#> 
#> --------------------------------------------------------------------------------