summarize_multi_group_model

summarize the results of a multi group model created with mxsem_group_by

Usage

summarize_multi_group_model(multi_group_model)

Arguments

multi_group_model

multi group model created with mxsem_group_by

Value

list with goup specific parameters and common parameters

Examples

# THE FOLLOWING EXAMPLE IS ADAPTED FROM
# https://openmx.ssri.psu.edu/docs/OpenMx/latest/_static/Rdoc/mxModel.html
library(mxsem)

model <- 'spatial =~ visual + cubes + paper
          verbal  =~ general + paragrap + sentence
          math    =~ numeric + series + arithmet'

mg_model <- mxsem(model = model,
                  data  = OpenMx::HS.ability.data) |>
  # we want separate models for all combinations of grades and schools:
  mxsem_group_by(grouping_variables = "school") |>
  mxTryHard()
#> Warning: This function is very experimental and may not yet work properly. Use with caution.
#> The following parameters will be the same across groups: 
#> The following parameters will be group specific: spatial→cubes, spatial→paper, verbal→paragrap, verbal→sentence, math→series, math→arithmet, visual↔visual, cubes↔cubes, paper↔paper, general↔general, paragrap↔paragrap, sentence↔sentence, numeric↔numeric, series↔series, arithmet↔arithmet, spatial↔spatial, spatial↔verbal, verbal↔verbal, spatial↔math, verbal↔math, math↔math, one→visual, one→cubes, one→paper, one→general, one→paragrap, one→sentence, one→numeric, one→series, one→arithmet
#> Running untitled40 with 60 parameters
#> 
#> Beginning initial fit attempt
#> Running untitled40 with 60 parameters
#> 
#>  Lowest minimum so far:  15969.1164364042
#>  Not all eigenvalues of the Hessian are positive: 402219341.891002, 24136125.6511393, 13629.6144673512, 13392.3232184624, 5863.79847475932, 4759.40902391292, 4360.27273268796, 1952.6181949062, 1825.43588373687, 1316.14320026041, 1149.69073403232, 357.020859626818, 139.587433865707, 71.9069260853621, 69.9404892014232, 46.269886249906, 42.6674235408757, 38.5891239225252, 24.7840310381383, 23.5787338184784, 22.843212497112, 21.5482242637395, 18.3282408584932, 15.9762169652866, 15.3060860161677, 13.9445602579177, 9.90204880539195, 6.10860281338663, 5.88811684225566, 5.50821258581185, 4.37657765716972, 4.36058933723348, 3.16509425753877, 2.38968859583729, 1.63544781213769, 1.58365930943974, 1.58124508589757, 1.38618515635273, 1.05930596548883, 0.739622567890229, 0.519989228275848, 0.479705953325449, 0.408746816359721, 0.367294348213997, 0.32782118531774, 0.18308465960063, 0.152239339155351, 0.0815597420870642, 0.0637842634380338, 0.0438074692131403, 0.0383380250844334, 0.0370122618603593, 0.0294527879632414, 0.0209545898869084, 0.0194967028842953, 0.00904037989428845, 0.00750007582366306, 0.00577013380997979, 2.1042732070824e-05, -0.000242048900633485
#> 
#> Beginning fit attempt 1 of at maximum 10 extra tries
#> Running untitled40 with 60 parameters
#>  Not all eigenvalues of the Hessian are positive: 402219341.891002, 24136125.6511393, 13629.6144673512, 13392.3232184624, 5863.79847475932, 4759.40902391292, 4360.27273268796, 1952.6181949062, 1825.43588373687, 1316.14320026041, 1149.69073403232, 357.020859626818, 139.587433865707, 71.9069260853621, 69.9404892014232, 46.269886249906, 42.6674235408757, 38.5891239225252, 24.7840310381383, 23.5787338184784, 22.843212497112, 21.5482242637395, 18.3282408584932, 15.9762169652866, 15.3060860161677, 13.9445602579177, 9.90204880539195, 6.10860281338663, 5.88811684225566, 5.50821258581185, 4.37657765716972, 4.36058933723348, 3.16509425753877, 2.38968859583729, 1.63544781213769, 1.58365930943974, 1.58124508589757, 1.38618515635273, 1.05930596548883, 0.739622567890229, 0.519989228275848, 0.479705953325449, 0.408746816359721, 0.367294348213997, 0.32782118531774, 0.18308465960063, 0.152239339155351, 0.0815597420870642, 0.0637842634380338, 0.0438074692131403, 0.0383380250844334, 0.0370122618603593, 0.0294527879632414, 0.0209545898869084, 0.0194967028842953, 0.00904037989428845, 0.00750007582366306, 0.00577013380997979, 2.1042732070824e-05, -0.000242048900633485
#> 
#> Beginning fit attempt 2 of at maximum 10 extra tries
#> Running untitled40 with 60 parameters
#> 
#>  Fit attempt worse than current best:  16034.292103261 vs 15969.1164364042
#> 
#> Beginning fit attempt 3 of at maximum 10 extra tries
#> Running untitled40 with 60 parameters
#> 
#>  Fit attempt generated errors
#> 
#> Beginning fit attempt 4 of at maximum 10 extra tries
#> Running untitled40 with 60 parameters
#> 
#>  Fit attempt generated errors
#> 
#> Beginning fit attempt 5 of at maximum 10 extra tries
#> Running untitled40 with 60 parameters
#> 
#>  Fit attempt generated errors
#> 
#> Beginning fit attempt 6 of at maximum 10 extra tries
#> Running untitled40 with 60 parameters
#> 
#>  Fit attempt worse than current best:  17594.0211372571 vs 15969.1164364042
#> 
#> Beginning fit attempt 7 of at maximum 10 extra tries
#> Running untitled40 with 60 parameters
#> 
#>  Lowest minimum so far:  15958.6071541106
#>  OpenMx status code 6 not in list of acceptable status codes, (0,0)
#>  Not all eigenvalues of the Hessian are positive: 2719378915.11693, 34800.2072628671, 11723.8036416556, 5721.03771791543, 4986.92947709393, 1950.94900882531, 1907.88896607652, 1177.6220159224, 367.650332082335, 145.852369663896, 71.9223227010296, 69.2515407861154, 46.2260556517324, 43.2340319301573, 42.4840601655504, 40.2432432552551, 24.8749582016622, 23.5446126388183, 22.6913050486704, 21.2909768863683, 17.3786025335836, 16.1418988551331, 15.8781745742528, 15.4196865643575, 14.53747095249, 10.8585528742527, 10.0852154363366, 6.85440858518139, 6.28057911989456, 5.87483150799385, 5.41766414393805, 4.52898868615475, 4.45752897219073, 3.16187337491505, 1.70723319033653, 1.63628256547188, 1.61610811861176, 1.58565358109331, 1.38594512868728, 1.06198438830775, 0.72335122326618, 0.525704917962496, 0.494468261292892, 0.469599529734287, 0.424092224535263, 0.336559705450283, 0.184882396604258, 0.166974970991784, 0.0829979567076024, 0.080670641186034, 0.0427431110978956, 0.040352738810421, 0.0389323961815644, 0.0286059956063712, 0.0209016180461412, 0.00870249668600295, 0.00786054847848456, 0.00277226041145968, 0.00020767587240249, -0.0937303690214503
#> 
#> Beginning fit attempt 8 of at maximum 10 extra tries
#> Running untitled40 with 60 parameters
#> 
#>  Lowest minimum so far:  15958.4903441559
#>  OpenMx status code 6 not in list of acceptable status codes, (0,0)
#>  Not all eigenvalues of the Hessian are positive: 2079320916.10271, 30396.7000669783, 10259.4439508966, 5713.13440690448, 4988.11397344534, 1947.70713503622, 1908.38422068654, 1172.69884195061, 365.536579830592, 145.802428484711, 71.9143656996535, 69.2555496511027, 46.2028930424893, 43.2243746887715, 42.4615495516083, 40.2426498691874, 24.8766778337418, 23.5024994707922, 22.691191805713, 21.2909906354227, 17.3784905181593, 16.1418048446003, 15.83508938806, 15.4150595392134, 14.5373028890394, 10.8613384078597, 9.6663475993521, 6.85439477977679, 6.27819351201922, 5.87722093661721, 5.4184878412359, 4.52893490567556, 4.36817602405602, 3.14510195141858, 1.70579306459845, 1.63428883273085, 1.61267426831974, 1.58572849548398, 1.38507244136374, 1.05499330141583, 0.719953910957078, 0.525519920823136, 0.494407585229849, 0.467979756489792, 0.424085410205642, 0.336397121325347, 0.183573948700779, 0.166196452449012, 0.0818615233182483, 0.0805410372218921, 0.042686757379403, 0.0389102850194145, 0.0360795117137158, 0.0285935630707645, 0.0192890150235191, 0.00867231813031054, 0.00783777470884716, 0.00266659480641782, -4.47050537845915e-05, -0.0970891847553774
#> 
#> Beginning fit attempt 9 of at maximum 10 extra tries
#> Running untitled40 with 60 parameters
#> 
#>  Lowest minimum so far:  15915.8789415703
#>  OpenMx status code 6 not in list of acceptable status codes, (0,0)
#> 
#> Beginning fit attempt 10 of at maximum 10 extra tries
#> Running untitled40 with 60 parameters
#> 
#>  Fit attempt worse than current best:  16336.1457566046 vs 15915.8789415703
#> 
#> Retry limit reached
#> 
#> 
#>  Retry limit reached; Best fit=15915.879 (started at 11279122)  (11 attempt(s): 8 valid, 3 errors)
#> 
#>  Uncertain solution found - consider parameter validity, try again, increase extraTries, change inits, change model, or check data!
#>  Start values from best fit:
#> 0.429709041690937,0.286030047787478,0.312238565941119,0.442337054882967,2.27699781494617,0.957027078451641,23.8911872492268,15.2671782093139,5.97877159249573,55.1571607513695,3.53182398114614,5.84648269350544,14.0477644022806,39.6605279012879,13.6894301730627,23.5792543032506,25.1601775244045,79.8355680779531,12.1044985286841,18.9595588642762,9.13222990696542,29.5793299430315,24.80006440603,14.3034508685122,44.8482308088995,9.95174136893617,18.8482850542459,14.4551756364226,18.7517565140437,25.8275638698641,0.318455519565273,0.222433538657814,0.302246251689723,0.479050993524961,-475.275921281042,-137.033055058914,17.4714718342727,20.7374851403738,6.42388486177301,48.0947544362934,3.66140621167075,6.68767797125442,19.4684988410478,12.506570232648,13.7481045102571,32.7961914094407,25.4042696529519,90.0685867419136,-0.0662101059570725,-0.0836064095583607,0.000288449888669922,29.6474749379358,23.9358581074698,14.1602741221549,36.6987601849222,8.46792550640095,15.9807704487346,13.8910286104286,17.5447806888955,22.7692557386679

# let's summarize the results:
summarize_multi_group_model(mg_model)
#> 
#> 
#> --------------------------------------------------------------------------------
#> Common Parameters:
#>  [1] name      matrix    row       col       Estimate  Std.Error lbound   
#>  [8] ubound    lboundMet uboundMet
#> <0 rows> (or 0-length row.names)
#> --------------------------------------------------------------------------------
#> group_1:
#> # A tibble: 1 × 1
#>   school     
#>   <fct>      
#> 1 Grant-White
#> 
#>                         name               matrix      row      col   Estimate
#> 1      spatial→cubes_group_1 untitled40_group_1.A    cubes  spatial  0.4297090
#> 2      spatial→paper_group_1 untitled40_group_1.A    paper  spatial  0.2860300
#> 3    verbal→paragrap_group_1 untitled40_group_1.A paragrap   verbal  0.3122386
#> 4    verbal→sentence_group_1 untitled40_group_1.A sentence   verbal  0.4423371
#> 5        math→series_group_1 untitled40_group_1.A   series     math  2.2769978
#> 6      math→arithmet_group_1 untitled40_group_1.A arithmet     math  0.9570271
#> 7      visual↔visual_group_1 untitled40_group_1.S   visual   visual 23.8911872
#> 8        cubes↔cubes_group_1 untitled40_group_1.S    cubes    cubes 15.2671782
#> 9        paper↔paper_group_1 untitled40_group_1.S    paper    paper  5.9787716
#> 10   general↔general_group_1 untitled40_group_1.S  general  general 55.1571608
#> 11 paragrap↔paragrap_group_1 untitled40_group_1.S paragrap paragrap  3.5318240
#> 12 sentence↔sentence_group_1 untitled40_group_1.S sentence sentence  5.8464827
#> 13   numeric↔numeric_group_1 untitled40_group_1.S  numeric  numeric 14.0477644
#> 14     series↔series_group_1 untitled40_group_1.S   series   series 39.6605279
#> 15 arithmet↔arithmet_group_1 untitled40_group_1.S arithmet arithmet 13.6894302
#> 16   spatial↔spatial_group_1 untitled40_group_1.S  spatial  spatial 23.5792543
#> 17    spatial↔verbal_group_1 untitled40_group_1.S  spatial   verbal 25.1601775
#> 18     verbal↔verbal_group_1 untitled40_group_1.S   verbal   verbal 79.8355681
#> 19      spatial↔math_group_1 untitled40_group_1.S  spatial     math 12.1044985
#> 20       verbal↔math_group_1 untitled40_group_1.S   verbal     math 18.9595589
#> 21         math↔math_group_1 untitled40_group_1.S     math     math  9.1322299
#> 22        one→visual_group_1 untitled40_group_1.M        1   visual 29.5793299
#> 23         one→cubes_group_1 untitled40_group_1.M        1    cubes 24.8000644
#> 24         one→paper_group_1 untitled40_group_1.M        1    paper 14.3034509
#> 25       one→general_group_1 untitled40_group_1.M        1  general 44.8482308
#> 26      one→paragrap_group_1 untitled40_group_1.M        1 paragrap  9.9517414
#> 27      one→sentence_group_1 untitled40_group_1.M        1 sentence 18.8482851
#> 28       one→numeric_group_1 untitled40_group_1.M        1  numeric 14.4551756
#> 29        one→series_group_1 untitled40_group_1.M        1   series 18.7517565
#> 30      one→arithmet_group_1 untitled40_group_1.M        1 arithmet 25.8275639
#>     Std.Error lbound ubound lboundMet uboundMet
#> 1  24.9030583     NA     NA     FALSE     FALSE
#> 2  39.6617636     NA     NA     FALSE     FALSE
#> 3  93.4319517     NA     NA     FALSE     FALSE
#> 4  70.1690300     NA     NA     FALSE     FALSE
#> 5   9.3347563     NA     NA     FALSE     FALSE
#> 6  16.7239047     NA     NA     FALSE     FALSE
#> 7   0.5212222  1e-06     NA     FALSE     FALSE
#> 8   1.0289159  1e-06     NA     FALSE     FALSE
#> 9   2.5982649  1e-06     NA     FALSE     FALSE
#> 10  0.2523245  1e-06     NA     FALSE     FALSE
#> 11  3.4808719  1e-06     NA     FALSE     FALSE
#> 12  1.9306669  1e-06     NA     FALSE     FALSE
#> 13  1.0522389  1e-06     NA     FALSE     FALSE
#> 14  0.3250950  1e-06     NA     FALSE     FALSE
#> 15  1.0898678  1e-06     NA     FALSE     FALSE
#> 16  0.6515880  1e-06     NA     FALSE     FALSE
#> 17  0.6048735     NA     NA     FALSE     FALSE
#> 18  0.2723814  1e-06     NA     FALSE     FALSE
#> 19  1.8694713     NA     NA     FALSE     FALSE
#> 20  1.2119350     NA     NA     FALSE     FALSE
#> 21  2.2908157  1e-06     NA     FALSE     FALSE
#> 22  4.2132696     NA     NA     FALSE     FALSE
#> 23  5.9194173     NA     NA     FALSE     FALSE
#> 24  9.4066757     NA     NA     FALSE     FALSE
#> 25  2.9316349     NA     NA     FALSE     FALSE
#> 26 10.8878331     NA     NA     FALSE     FALSE
#> 27  8.1087667     NA     NA     FALSE     FALSE
#> 28  5.9854404     NA     NA     FALSE     FALSE
#> 29  3.3273373     NA     NA     FALSE     FALSE
#> 30  6.0909945     NA     NA     FALSE     FALSE
#> --------------------------------------------------------------------------------
#> group_2:
#> # A tibble: 1 × 1
#>   school 
#>   <fct>  
#> 1 Pasteur
#> 
#>                         name               matrix      row      col
#> 31     spatial→cubes_group_2 untitled40_group_2.A    cubes  spatial
#> 32     spatial→paper_group_2 untitled40_group_2.A    paper  spatial
#> 33   verbal→paragrap_group_2 untitled40_group_2.A paragrap   verbal
#> 34   verbal→sentence_group_2 untitled40_group_2.A sentence   verbal
#> 35       math→series_group_2 untitled40_group_2.A   series     math
#> 36     math→arithmet_group_2 untitled40_group_2.A arithmet     math
#> 37     visual↔visual_group_2 untitled40_group_2.S   visual   visual
#> 38       cubes↔cubes_group_2 untitled40_group_2.S    cubes    cubes
#> 39       paper↔paper_group_2 untitled40_group_2.S    paper    paper
#> 40   general↔general_group_2 untitled40_group_2.S  general  general
#> 41 paragrap↔paragrap_group_2 untitled40_group_2.S paragrap paragrap
#> 42 sentence↔sentence_group_2 untitled40_group_2.S sentence sentence
#> 43   numeric↔numeric_group_2 untitled40_group_2.S  numeric  numeric
#> 44     series↔series_group_2 untitled40_group_2.S   series   series
#> 45 arithmet↔arithmet_group_2 untitled40_group_2.S arithmet arithmet
#> 46   spatial↔spatial_group_2 untitled40_group_2.S  spatial  spatial
#> 47    spatial↔verbal_group_2 untitled40_group_2.S  spatial   verbal
#> 48     verbal↔verbal_group_2 untitled40_group_2.S   verbal   verbal
#> 49      spatial↔math_group_2 untitled40_group_2.S  spatial     math
#> 50       verbal↔math_group_2 untitled40_group_2.S   verbal     math
#> 51         math↔math_group_2 untitled40_group_2.S     math     math
#> 52        one→visual_group_2 untitled40_group_2.M        1   visual
#> 53         one→cubes_group_2 untitled40_group_2.M        1    cubes
#> 54         one→paper_group_2 untitled40_group_2.M        1    paper
#> 55       one→general_group_2 untitled40_group_2.M        1  general
#> 56      one→paragrap_group_2 untitled40_group_2.M        1 paragrap
#> 57      one→sentence_group_2 untitled40_group_2.M        1 sentence
#> 58       one→numeric_group_2 untitled40_group_2.M        1  numeric
#> 59        one→series_group_2 untitled40_group_2.M        1   series
#> 60      one→arithmet_group_2 untitled40_group_2.M        1 arithmet
#>         Estimate    Std.Error lbound ubound lboundMet uboundMet
#> 31  3.184555e-01 2.685787e+01     NA     NA     FALSE     FALSE
#> 32  2.224335e-01 4.787292e+01     NA     NA     FALSE     FALSE
#> 33  3.022463e-01 1.022219e+02     NA     NA     FALSE     FALSE
#> 34  4.790510e-01 7.132214e+01     NA     NA     FALSE     FALSE
#> 35 -4.752759e+02 7.055605e-02     NA     NA     FALSE     FALSE
#> 36 -1.370331e+02 1.038800e-01     NA     NA     FALSE     FALSE
#> 37  1.747147e+01 5.397463e-01  1e-06     NA     FALSE     FALSE
#> 38  2.073749e+01 8.179268e-01  1e-06     NA     FALSE     FALSE
#> 39  6.423885e+00 2.577158e+00  1e-06     NA     FALSE     FALSE
#> 40  4.809475e+01 2.865727e-01  1e-06     NA     FALSE     FALSE
#> 41  3.661406e+00 3.556834e+00  1e-06     NA     FALSE     FALSE
#> 42  6.687678e+00 1.686791e+00  1e-06     NA     FALSE     FALSE
#> 43  1.946850e+01 9.076268e-01  1e-06     NA     FALSE     FALSE
#> 44  1.250657e+01 4.059720e-01  1e-06     NA     FALSE     FALSE
#> 45  1.374810e+01 1.215618e+00  1e-06     NA     FALSE     FALSE
#> 46  3.279619e+01 5.320454e-01  1e-06     NA     FALSE     FALSE
#> 47  2.540427e+01 4.930934e-01     NA     NA     FALSE     FALSE
#> 48  9.006859e+01 2.218613e-01  1e-06     NA     FALSE     FALSE
#> 49 -6.621011e-02 3.179827e+02     NA     NA     FALSE     FALSE
#> 50 -8.360641e-02 1.981168e+02     NA     NA     FALSE     FALSE
#> 51  2.884499e-04 8.005716e+04  1e-06     NA      TRUE     FALSE
#> 52  2.964747e+01 4.366445e+00     NA     NA     FALSE     FALSE
#> 53  2.393586e+01 5.375120e+00     NA     NA     FALSE     FALSE
#> 54  1.416027e+01 9.541702e+00     NA     NA     FALSE     FALSE
#> 55  3.669876e+01 3.183354e+00     NA     NA     FALSE     FALSE
#> 56  8.467926e+00 1.120941e+01     NA     NA     FALSE     FALSE
#> 57  1.598077e+01 7.719504e+00     NA     NA     FALSE     FALSE
#> 58  1.389103e+01 5.661442e+00     NA     NA     FALSE     FALSE
#> 59  1.754478e+01 3.782753e+00     NA     NA     FALSE     FALSE
#> 60  2.276926e+01 6.552870e+00     NA     NA     FALSE     FALSE
#> 
#> --------------------------------------------------------------------------------